### 2019-08-14 #42 @whedon
[PRE REVIEW]: TSML (Time Series Machine Learning)
**Submitting author:** @ppalmes (<a href="http://orcid.org/0000-0002-3145-6356">Paulito Palmes</a>)
**Repository:** <a href="https://github.com/IBM/TSML.jl" target ="_blank">https://github.com/IBM/TSML.jl/tree/master/paper</a>

Thanks for submitting your paper to JuliaCon @ppalmes. The JuliaCon editor @matbesancon, will work with you on this issue to find a reviewer for your submission before creating the main review issue.

@ppalmes if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](http://bit.ly/2V9GSFq) have already agreed to review for JuliaCon and may be suitable for this submission.

The JuliaCon submission bot @whedon is here to help you find and assign reviewers and start the main review. To find out what @whedon can do for you type:
