### 2019-08-14 5d4dea3
update travis to use julia 1.1
