### 2019-08-20 #42 @matbesancon
Hi @ppalmes, the header.tex file is not supposed to be touched, it's a machine-generated file. Instead, you should fill the information in the `paper.yml` file
