### 2019-08-20 5aa6975
remove unnecessary package and edit Project.toml
