+++ src/outliernicer.jl
@@ -2,7 +2,6 @@
 
 using Dates
 using DataFrames
-using Random
 using Statistics
 using StatsBase: iqr, quantile, sample
 
