### 2019-08-20 #42 @ppalmes
Thanks. I'll update based on your comments.

On Tue, Aug 20, 2019 at 9:22 AM Mathieu Besançon <notifications@github.com>
wrote:

> Hi @ppalmes <https://github.com/ppalmes>, the header.tex file is not
> supposed to be touched, it's a machine-generated file. Instead, you should
> fill the information in the paper.yml file
>
