+++ v2.3.7/siteinfo.js
@@ -1,0 +1,1 @@
+var DOCUMENTER_CURRENT_VERSION = "v2.3.7";
