### 2019-08-26 #42 @ppalmes
hi @matbesancon, i need your help. i would like to move out my paper from repo outside the directory of my TSML package. i'm having issues with *chmod* bug in TSML which I suspect can be caused by having the paper subdirectory. I'd like to move my paper repo into: https://github.com/ppalmes/TSMLJuliaCon2019Paper. Do you know how to change the pointer to update the location?
