### 2019-08-26 #42 @matbesancon
The paper folder should not influence the package at all. And you can change chmod for the files inside. From what I see in the repo:
1. you need to change paper.yml to the corresponding data, you should not use header.tex
2. you should not comment out `%\input{bib.tex}`, this is likely causing the error
