+++ paper/header.tex
@@ -1,11 +1,0 @@
-% **************GENERATED FILE, DO NOT EDIT**************
-
-\title{TSML (Time Series Machine Learning)}
-
-\author{Paulito P. Palmes}
-\author{Joern Ploennigs}
-\author{Niall Brady}
-\affil{IBM Dublin Research Lab}
-
-\keywords{Julia, Time Series, Machine Learning, Filters, Feature Extraction, Classification, Prediction, Aggregation, Imputation}
-
