+++ paper/paper.tex
@@ -294,10 +294,12 @@ end
 
 TODO: 
 \begin{itemize}
-\item add plots
-\item add stats output
-\item add more examples from docs
+\item add plots and tables
+\item add threading examples for scalability (buggy)
+\item add simulations of different datasets to show general applicability (still doing simulations)
+\item add results of simulations: Earthquakes sensor; Electric Devices; car sensors; refrigeration devices; household consumption
 \end{itemize}
+
 %\begin{verbatim}
 %\bibliographystyle{juliacon}
 %\bibliography{ref}
@@ -306,7 +308,7 @@ TODO:
 %parties, the ref.bib file should be sent with it.
 %\cite{bezanson2017julia}
 
-%\input{bib.tex}
+\input{bib.tex}
 
 \end{document}
 
