### 2019-08-27 #42 @ppalmes
hi @matbesancon, tried to follow your suggestion, still getting some errors.
