### 2019-08-27 31d6859
paper updates
