### 2019-08-27 a7a8bcf
paper updates
