+++ paper/paper.yml
@@ -11,6 +11,7 @@ keywords:
   - Imputation
 authors:
   - name: Paulito Palmes
+    orcid: 0000-0002-3145-6356
     affiliation: 1
   - name: Joern Ploennigs
     affiliation: 1
