### 2019-08-27 #42 @ppalmes
@matbesancon, i merged your PR as well as added my orcid. still the same issue.
