### 2019-08-27 #42 @ppalmes
oh wow. thanks. what should i avoid to make sure i won't have this issue?
