### 2019-08-27 #42 @matbesancon
it was in the submission repo, the whole `/tree/master/paper` path was not needed
