### 2019-08-27 #42 @matbesancon
@christianpeel would you be available to review this paper?
