### 2019-08-27 #42 @ppalmes
ah thanks. regarding https://github.com/JuliaLang/Pkg.jl/issues/1313, i hope the paper subdirectory which i added recently did not cause this bug.
