### 2019-08-27 #42 @ppalmes
@christianpeel, i think it's ok because the target of the package is for general consumption and high level usage. it's more of combo of package tutorial, demo, and collections of real applications avoiding technical details on ML implementations.
