+++ paper/paper.tex
@@ -147,22 +147,111 @@ of missing blocks count is now empty.
 
 \vskip 3pt
 
-We can also visualise our time series data using the \emph{plotter} filter instead of the \emph{statifier} as show below:
+We can also visualise our time series data using the \emph{plotter} filter instead of the \emph{statifier} as show in Fig~\ref{fig:mplot}. Looking closely, you will see discontinuities in the plot due to blocks of missing data.
 
 \begin{lstlisting}
 pltr=Plotter(Dict(:interactive => true))
 plpipeline = Pipeline(Dict(
    :transformers => [csvfilter, valgator, pltr]))
-fit!(plpipeline)
-myplot = transform!(plpipeline)
+fit!(plpipeline); transform!(plpipeline)
 \end{lstlisting}
 
 \begin{figure}[htbp]
    \centering
    \includegraphics[width=\columnwidth]{mplot.png} % requires the graphicx package
-   \caption{example caption}
-   \label{fig:example}
+   \caption{Plot with Missing Data}
+   \label{fig:mplot}
 \end{figure}
+
+Using the imputation pipeline described before, we can visualise the result of imputation by replacing the Statifier with the Plotter filter. Figure~\ref{fig:amplot} shows the plot after imputation which gets rid of missing data.
+
+\begin{lstlisting}[language = Julia]
+bplpipeline = Pipeline(Dict(
+  :transformers => [csvfilter, valgator, 
+                    valnner,pltr]))
+fit!(bplpipeline); transform!(bplpipeline)
+\end{lstlisting}
+
+\begin{figure}[htbp]
+   \centering
+   \includegraphics[width=\columnwidth]{amplot.png} % requires the graphicx package
+   \caption{Plot after Data Imputation}
+   \label{fig:amplot}
+\end{figure}
+
+\section{Processing Monotonic Time Series}
+In this section, we will show additional filters we use to handle monotonic data which are commonly used in energy/water meter sensors as well as in footfall sensors. In the former case, the time series type is strictly monotonically increasing while in the latter case, the monotonicity happens daily by resetting the sensor to certain baseline. We use the filter called \emph{Monotonicer} which automatically detects the two types of monotonic sensors and apply the normalisation accordingly. 
+
+\begin{lstlisting}[language = Julia]
+mono = joinpath(dirname(pathof(TSML)),
+     "../data/typedetection/monotonic.csv")
+dailymono = joinpath(dirname(pathof(TSML)),
+     "../data/typedetection/dailymonotonic.csv")
+ 
+monocsv = CSVDateValReader(Dict(
+     :filename=>mono,
+     :dateformat=>"dd/mm/yyyy HH:MM"))
+dailymonocsv = CSVDateValReader(Dict(
+     :filename=>dailymono,
+     :dateformat=>"dd/mm/yyyy HH:MM"))
+\end{lstlisting}
+
+Let us plot the monotonic data with the usual workflow of aggregating and imputing the data first.
+
+\begin{lstlisting}[language = Julia]
+monopipeline = Pipeline(Dict(
+  :transformers => [monofilecsv,valgator,
+                    valputer,pltr]))
+fit!(monopipeline); transform!(monopipeline)
+\end{lstlisting}
+
+\begin{figure}[htbp]
+   \centering
+   \includegraphics[width=0.9\columnwidth]{mono.png} % requires the graphicx package
+   \caption{Monotonic Time Series}
+   \label{fig:mono}
+\end{figure}
+
+Let us now normalise the data by adding the \emph{Monotonicer} filter in the pipeline and plot.
+
+\begin{lstlisting}[language = Julia]
+mononicer = Monotonicer(Dict())
+
+monopipeline = Pipeline(Dict(
+  :transformers => [monofilecsv,valgator,valputer,
[…]
