### 2019-08-27 5069eb4
remove index affiliation
