+++ paper/paper.tex
@@ -249,7 +249,7 @@ fit!(monopipeline); transform!(monopipeline)
    \label{fig:outnicer}
 \end{figure}
 
-We follow similar workflow of normalizing daily monotonic time series. We visualise first the original data, apply the monotonic and outlier, and plot the cleaned data.
+We follow similar workflow of normalizing daily monotonic time series. We visualise first the original data, apply the monotonic and outlier filters, and plot the cleaned data to make sure monotonicity is gone.
 
 
 
