### 2019-08-27 #42 @ppalmes
hi @matbesancon, you can by the way add me to review papers in machine learning, time series, neural networks, data mining, biomedical engineering, internet of things
