### 2019-09-03 #42 @matbesancon
@MikeInnes would you be able to review this paper?
