+++ paper/paper.tex
@@ -23,18 +23,18 @@
 
 \section{Introduction}
 
-\textbf{TSML} is a package for time series data processing, classification, and prediction. 
+\textbf{TSML} is a Julia\cite{bezanson2017julia} package for time series data processing, classification, and prediction. 
 It provides common API for ML libraries from Python's ScikitLearn, 
 R's caret, and native Julia MLs for seamless integration of heterogeneous 
 libraries to create complex ensembles for robust time-series prediction, clustering, and classification.
 TSML has the following features:
-\begin{itemize}
+\begin{enumerate}
 \item data type clustering/classification for automatic data discovery
 \item aggregation based on date/time interval
 \item imputation based on symmetric Nearest Neighbors
-\item statistical metrics for data quality assessment
+\item statistical metrics for data quality assessment and classification input features
 \item ML wrapper with more than 100+ libraries from caret, scikitlearn, and julia
-\item date/value matrix conversion of 1-D time series using sliding windows for ML input
+\item date/value matrix conversion of 1-D time series using sliding windows to generate features for ML prediction
 \item pipeline API for high-level description of the processing workflow
 \item specific cleaning/normalization workflow based on data type
 \item automatic selection of optimised ML model
@@ -42,7 +42,7 @@ TSML has the following features:
 \item extensible architecture using just two main interfaces: fit and transform
 \item meta-ensembles for automatic feature and model selection
 \item support for distributed computation for scalability and speed
-\end{itemize}
+\end{enumerate}
 
 The \textbf{TSML} package assumes a two-column input composed of \emph{dates} and \emph{values}. 
 The first part of the workflow aggregates values based on the specified date/time 
@@ -80,14 +80,14 @@ assessment.
 
 \begin{lstlisting}[language = Julia]
 fname = joinpath(dirname(pathof(TSML)),
-   "../data/testdata.csv")
+  "../data/testdata.csv")
 csvfilter = CSVDateValReader(Dict(
-   :filename=>fname,
-   :dateformat=>"dd/mm/yyyy HH:MM"))
+  :filename=>fname,
+  :dateformat=>"dd/mm/yyyy HH:MM"))
 valgator = DateValgator(Dict(
-   :dateinterval=>Dates.Hour(1)))
-valnner = DateValNNer(Dict(
-   :dateinterval=>Dates.Hour(1)))
+  :dateinterval=>Dates.Hour(1)))
+valputer = DateValNNer(Dict(
+  :dateinterval=>Dates.Hour(1)))
 stfier = Statifier(Dict(:processmissing=>true))
 \end{lstlisting}
 
@@ -97,7 +97,7 @@ by aggregating the time series hourly and check the data quality using the
 
 \begin{lstlisting}[language = Julia]
 apipeline = Pipeline(Dict(
-   :transformers => [csvfilter, valgator, stfier]))
+  :transformers => [csvfilter, valgator, stfier]))
 fit!(apipeline)
 mystats = transform!(apipeline)
 @show mystats
@@ -119,6 +119,8 @@ and understanding of the entire workflow. In addition, extending TSML functional
 matter of creating a new data type filter and define its own  \emph{fit} and \emph{transform} 
 functions.
 
+\vskip 3pt
+
 In the \emph{Statifier} filter result, blocks of missing data is indicated by column names starting
 with \emph{b}. Running the code above indicates that there are plenty of missing data blocks.
[…]
