### 2019-09-04 d869b2e
add citations
