+++ paper/paper.tex
@@ -68,8 +68,8 @@ There are more than hundred classifiers and regression functions available using
 \section{TSML Workflow}
 \label{sec:tsmlworkflow}
 %
-
-\textbf{TSML} workflow borrows the idea of Unix pipeline.
+ Did
+\textbf{TSML} workflow borrows the idea of Unix pipeline\cite{orchestra2014, combineml2016}.
 The main elements in a pipeline are series of filters
 each performing one specific task and does it well. 
 
