### 2019-09-04 #42 @ppalmes
@matbesancon, hi mat. can you move the repo pointer to https://github.com/ppalmes/TSMLJuliaCon2019Paper instead of the current one? I would like to move it out of my TSML code director to separate the paper edit commits from my code commits. thanks.
