### 2019-09-04 #42 @matbesancon
It's encouraged to have the paper at the same place as the code for people to review.
One thing people have done is:
1. Place the paper in a special `juliacon-paper` branch
2. Remove it from master
3. Build with whedon as `@whedon generate pdf from branch my-branch`
