+++ paper/paper.tex
@@ -1,555 +1,0 @@
-
-% JuliaCon proceedings template
-\documentclass{juliacon}
-\setcounter{page}{1}
-
-\begin{document}
-
-\input{header}
-
-\maketitle
-
-\begin{abstract}
-
-  Over the past years, the industrial sector has seen many innovations brought about by automation. 
-  Inherent in this automation is the installation of sensor networks for status monitoring and data collection. 
-  One of the major challenges in these data-rich environments is how to extract and exploit information from 
-  these large volume of data to detect anomalies, discover patterns to reduce downtimes and manufacturing 
-  errors, reduce energy usage, predict faults/failures, effective maintenance schedules, etc. 
-  To address these issues, we developed TSML. Its technology is based 
-  on using the pipeline of lightweight filters as building blocks to process huge amount of industrial time series data in parallel.  
-
-\end{abstract}
-
-\section{Introduction}
-
-\textbf{TSML} is a Julia\cite{bezanson2017julia} package for time series data processing, classification, and prediction. 
-It provides common API for ML libraries from Python's ScikitLearn, 
-R's caret, and native Julia MLs for seamless integration of heterogeneous 
-libraries to create complex ensembles for robust time-series prediction, clustering, and classification.
-TSML has the following features:
-\begin{enumerate}
-\item data type clustering/classification for automatic data discovery
-\item aggregation based on date/time interval
-\item imputation based on symmetric Nearest Neighbors
-\item statistical metrics for data quality assessment and classification input features
-\item ML wrapper with more than 100+ libraries from caret, scikitlearn, and julia
-\item date/value matrix conversion of 1-D time series using sliding windows to generate features for ML prediction
-\item pipeline API for high-level description of the processing workflow
-\item specific cleaning/normalization workflow based on data type
-\item automatic selection of optimised ML model
-\item automatic segmentation of time-series data into matrix form for ML training and prediction 
-\item extensible architecture using just two main interfaces: fit and transform
-\item meta-ensembles for automatic feature and model selection
-\item support for distributed computation for scalability and speed
-\end{enumerate}
-
-The \textbf{TSML} package assumes a two-column input composed of \emph{dates} and \emph{values}. 
-The first part of the workflow aggregates values based on the specified date/time 
-interval which minimizes occurrence of missing values and noise. The aggregated 
-data is then left-joined to the complete sequence of dates in a specified date/time interval. 
-Remaining missing values are replaced by \textit{k} nearest neighbors where \textit{k} is the symmetric 
-distance from the location of missing value. This approach can be called several 
-times until there are no more missing values.
-
-For prediction tasks, TSML extracts the date features and 
-convert the value column into matrix form parameterized by 
-the size and stride of the sliding window. The final part joins
- the date features and the value matrix to serve as input to the 
- ML with the output representing the values of the time periods 
- to be predicted ahead of time.
- 
-\textbf{TSML} uses a pipeline which iteratively calls the \emph{fit} and \emph{transform}
-families of functions relying on `multiple dispatch` to select the correct 
-algorithm from the steps outlined above. Machine learning functions in 
[…]
