+++ paper/ref.bib
@@ -1,60 +1,0 @@
-%% This BibTeX bibliography file was created using BibDesk.
-%% https://bibdesk.sourceforge.io/
-
-%% Created for Paulito Palmes at 2019-09-04 10:36:20 +0100 
-
-
-%% Saved with string encoding Unicode (UTF-8) 
-
-
-
-@article{bezanson2017julia,
-	Author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
-	Journal = {SIAM review},
-	Number = {1},
-	Pages = {65--98},
-	Publisher = {SIAM},
-	Title = {Julia: A fresh approach to numerical computing},
-	Volume = {59},
-	Year = {2017}}
-
-@article{Moore1990,
-	Author = {Moore, James T. and Bard, Jonathan F.},
-	Doi = {10.1287/opre.38.5.911},
-	Issn = {0030-364X},
-	Journal = {Operations Research},
-	Keywords = {games: noncooperative,integer: branch-and-bound algorithms,programming},
-	Mendeley-Groups = {bilevel - hierarchical},
-	Month = {oct},
-	Number = {5},
-	Pages = {911--921},
-	Publisher = {INFORMS},
-	Title = {{The Mixed Integer Linear Bilevel Programming Problem}},
-	Url = {http://pubsonline.informs.org/doi/abs/10.1287/opre.38.5.911},
-	Volume = {38},
-	Year = {1990},
-	Bdsk-Url-1 = {http://pubsonline.informs.org/doi/abs/10.1287/opre.38.5.911},
-	Bdsk-Url-2 = {https://doi.org/10.1287/opre.38.5.911}}
-
-@misc{orchestra2014,
-  author = {Jenkins, S.},
-  title = {Orchestra.jl},
-  year = {2014},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  howpublished = {\url{https://github.com/svs14/Orchestra.jl}},
-  commit = {1d01e4b94e08fa19ec0ce41f0d0a1bdc4516bd4f}
-}
-
-
-@misc{combineml2016,
-  author = {Palmes, P.},
-  doi = {10.5281/zenodo.3385904},
-  url = {https://zenodo.org/badge/latestdoi/83453995},
-  title = {{CombineML}: A Package to Create Heterogeneous Ensembles of {ML} from {ScikitLearn}, {Caret}, and {Julia}},
-  year = {2016},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  howpublished = {\url{https://github.com/ppalmes/CombineML.jl}},
-  commit = {a94fd31f15df1a7eff1e1fdc2a067c67666c3654}
-}
