+++ paper/paper.yml
@@ -1,20 +1,0 @@
-title: "TSML (Time Series Machine Learning)"
-keywords:
-  - Julia
-  - Time Series
-  - Machine Learning
-  - Filters
-  - Feature Extraction
-  - Classification
-  - Prediction
-  - Aggregation
-  - Imputation
-authors:
-  - name: Paulito Palmes
-    orcid: 0000-0002-3145-6356
-  - name: Joern Ploennigs
-  - name: Niall Brady
-affiliations:
-  - name: IBM Dublin Research Lab
-date: 27 August 2019
-bibliography: ref.bib
