+++ .travis.yml
@@ -34,12 +34,12 @@ jobs:
       name: "HTML"
       after_success: skip
       services: docker
-   # - stage: "PDF Documentation"
-   #   julia: 1.1
-   #   os: osx
-   #   script:
-   #     - julia --project=docs/pdf/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
-   #     - julia --project=docs/pdf/ docs/pdf/make.jl
-   #   name: "PDF"
-   #   after_success: skip
-   #   services: docker
+    - stage: "PDF Documentation"
+      julia: 1.1
+      os: osx
+      script:
+        - julia --project=docs/pdf/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
+        - julia --project=docs/pdf/ docs/pdf/make.jl
+      name: "PDF"
+      after_success: skip
+      services: docker
