### 2019-09-05 #42 @matbesancon
@theogf would you be available to review this paper?
