### 2019-09-05 f5f6279
remove zepdoctor aliases
