+++ v2.3.3/siteinfo.js
@@ -1,1 +1,0 @@
-var DOCUMENTER_CURRENT_VERSION = "v2.3.3";
