### 2019-09-06 #51 @whedon
[REVIEW]: TSML (Time Series Machine Learning)
**Submitting author:** @ppalmes (<a href="http://orcid.org/0000-0002-3145-6356">Paulito Palmes</a>)
**Repository:** <a href="https://github.com/IBM/TSML.jl" target ="_blank">https://github.com/IBM/TSML.jl</a>
