+++ v2.3.8/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"man/aggregation/#","page":"Aggregation","title":"Aggregation","text":"Author = \"Paulito P. Palmes\"","category":"page"},{"location":"man/aggregation/#Aggregation-1","page":"Aggregation","title":"Aggregation","text":"","category":"section"},{"location":"man/aggregation/#","page":"Aggregation","title":"Aggregation","text":"DateValgator is a data type that supports operation  for aggregation to minimize noise and  lessen the occurrence of missing data. It expects to receive one argument which is the date-time interval for grouping values by taking  their median. For example, hourly median as the basis of aggregation can be carried out by passing this argument: :dateinterval => Dates.Hour(1)","category":"page"},{"location":"man/aggregation/#","page":"Aggregation","title":"Aggregation","text":"To illustrate DateValgator usage, let's start by  generating an artificial data with sample frequencey every 5 minutes and print the first 10 rows.","category":"page"},{"location":"man/aggregation/#","page":"Aggregation","title":"Aggregation","text":"using TSML\n\ngdate = DateTime(2014,1,1):Dates.Minute(5):DateTime(2014,5,1)\ngval = rand(length(gdate))\ndf = DataFrame(Date=gdate,Value=gval)\nnothing #hide","category":"page"},{"location":"man/aggregation/#","page":"Aggregation","title":"Aggregation","text":"first(df,10)","category":"page"},{"location":"man/aggregation/#DateValgator-1","page":"Aggregation","title":"DateValgator","text":"","category":"section"},{"location":"man/aggregation/#","page":"Aggregation","title":"Aggregation","text":"Let's apply the aggregator and try diffent groupings: hourly vs half hourly vs daily aggregates of the data.","category":"page"},{"location":"man/aggregation/#","page":"Aggregation","title":"Aggregation","text":"using TSML\n\nhourlyagg = DateValgator(Dict(:dateinterval => Dates.Hour(1)))\nhalfhourlyagg = DateValgator(Dict(:dateinterval => Dates.Minute(30)))\ndailyagg = DateValgator(Dict(:dateinterval => Dates.Day(1)))\n\nfit!(halfhourlyagg,df)\nhalfhourlyres = transform!(halfhourlyagg,df)\n\nfit!(hourlyagg,df)\nhourlyres = transform!(hourlyagg,df)\n\nfit!(dailyagg,df)\ndailyres = transform!(dailyagg,df)\nnothing #hide","category":"page"},{"location":"man/aggregation/#","page":"Aggregation","title":"Aggregation","text":"The first 5 rows of half-hourly, hourly, and daily aggregates:","category":"page"},{"location":"man/aggregation/#","page":"Aggregation","title":"Aggregation","text":"first(halfhourlyres,5)\nfirst(hourlyres,5)\nfirst(dailyres,5)","category":"page"},{"location":"tutorial/monotonic_plotting/#","page":"Monotonic Detection and Plotting","title":"Monotonic Detection and Plotting","text":"Author = \"Paulito P. Palmes\"","category":"page"},{"location":"tutorial/monotonic_plotting/#Monotonic-Detection-and-Plotting-1","page":"Monotonic Detection and Plotting","title":"Monotonic Detection and Plotting","text":"","category":"section"},{"location":"tutorial/monotonic_plotting/#","page":"Monotonic Detection and Plotting","title":"Monotonic Detection and Plotting","text":"One important preprocessing step for time series data processing is the detection  of monotonic data and transform it to non-monotonic type by using the finite difference operator.","category":"page"},{"location":"tutorial/monotonic_plotting/#Artificial-Data-Example-1","page":"Monotonic Detection and Plotting","title":"Artificial 
[…]
