+++ versions.js
@@ -1,0 +1,5 @@
+var DOC_VERSIONS = [
+  "stable",
+  "v2.3",
+  "dev",
+];
