### 2019-09-08 58fa461
removing strange symlinks
