+++ versions.js
@@ -1,5 +1,0 @@
-var DOC_VERSIONS = [
-  "stable",
-  "v2.3",
-  "dev",
-];
