+++ paper/paper.tex
@@ -384,7 +384,7 @@ different languages. For illustration purposes, we will show how to
 ensemble ML libraries from ScikitLearn and Caret using \textbf{TSML} 
 meta-ensembles that support the \emph{fit} and \emph{transform} APIs.
 
-\vskip 6pt
+\subsection{Parallel TSML Using Distributed Workflow}
 
 We will use Julia's built-in support for parallelism by using the \emph{Distributed} library. We also let Julia detect the number of processors available and activate them.
 
@@ -481,7 +481,7 @@ Next, we setup the pipeline for training and prediction inside the \emph{predict
 end
 \end{lstlisting}
 
-Finally, we setup the \emph{parallel model} function to run different learners distributed to different workers running in parallel relying on Julia's native support of parallelism.
+Finally, we setup the \emph{parallel model} function to run different learners distributed to different workers running in parallel relying on Julia's native support of parallelism. Take note that there are two parallelisms in the code. The first one is the distribution of task in different trials and the second one is the distribution of tasks among different models. It is interesting to note that with this relative compact function definition, the Julia language makes it easy to define a parallel task within another parallel task without any problem.
 
 \begin{lstlisting}[language = Julia]
 function parallelmodel(learners::Dict,
@@ -523,10 +523,13 @@ learners=Dict(
   :skl_knn=>sk_knn,:skl_gb=>sk_gb,
   :skl_extratree=>sk_extratree, :sk_rf=>sk_rf
 )
-df = parallelmodel(learners,tsdata;trials=3)
-\end{lstlisting}
 
-Figure~\ref{fig:sim}  shows a snapshot of running workers exploiting the distributed library of Julia. There are 8 workers running in parallel over 12 different machine learning classifiers.
+datadir = joinpath("tsdata/")
+tsdata = extract_features_from_timeseries(datadir)
+first(tsdata,5)
+
+respar = parallelmodel(learners,tsdata;trials=3)
+\end{lstlisting}
 
 \begin{figure}[htbp]
    \centering
@@ -538,15 +541,116 @@ Figure~\ref{fig:sim}  shows a snapshot of running workers exploiting the distrib
 \begin{figure}[htbp]
    \centering
    \includegraphics[width=0.6\columnwidth]{results.png} % requires the graphicx package
-   \caption{Classification Performance in 3 Trials}
+   \caption{@distributed: Classification Performance in 3 Trials}
    \label{fig:performance}
 \end{figure}
 
-TODO: 
-\begin{itemize}
-\item add threading examples for scalability (buggy, need fixes)
-\item add simulations of different datasets to show general applicability (still doing simulations): Earthquakes sensor; Electric Devices; car sensors; refrigeration devices; household consumption
-\end{itemize}
+Figures~\ref{fig:sim}~and~\ref{fig:performance}  shows a snapshot of running workers exploiting the distributed library of Julia and the classification performance of each model, respectively. There are 8 workers running in parallel over 12 different machine learning classifiers. From the results, \emph{ExtraTree} from \emph{ScikitLearn} has the best performance with around 91.67\% accuracy followed by \emph{TreeBag} and \emph{SVMLinear} from \emph{Caret} library with 83.33 \% accuracy for both. With this workflow, it becomes trivial to search for optimal model by running them in parallel relying on \emph{Julia} to do the low-level tasks of scheduling, queueing, etc and making sure that compute resources are fairly optimised.
+
+\subsection{Parallel TSML Using Threads Workflow}
[…]
