### 2019-09-10 5b53d8f
minor edits for emphasis
