+++ paper/paper.tex
@@ -3,6 +3,9 @@
 \documentclass{juliacon}
 \setcounter{page}{1}
 
+\hyphenation{TSMLextra}
+\hyphenation{TSML}
+
 \begin{document}
 
 \input{header}
@@ -16,24 +19,24 @@
   One of the major challenges in these data-rich environments is how to extract and exploit information from 
   these large volume of data to detect anomalies, discover patterns to reduce downtimes and manufacturing 
   errors, reduce energy usage, predict faults/failures, effective maintenance schedules, etc. 
-  To address these issues, we developed TSML. Its technology is based 
+  To address these issues, we developed \textbf{TSML}. Its technology is based 
   on using the pipeline of lightweight filters as building blocks to process huge amount of industrial time series data in parallel.  
 
 \end{abstract}
 
 \section{Introduction}
 
-\textbf{TSML} is a Julia\cite{bezanson2017julia} package for time series data processing, classification, and prediction. 
-It provides common API for ML libraries from Python's ScikitLearn, 
-R's caret, and native Julia MLs for seamless integration of heterogeneous 
-libraries to create complex ensembles for robust time-series prediction, clustering, and classification.
-TSML has the following features:
+\textbf{TSML}\cite{tsml2019} is a \emph{Julia}\cite{bezanson2017julia} package for time series data processing, classification, and prediction. 
+It provides common API for ML libraries from Python's \emph{Scikit-Learn}, 
+R's \emph{Caret}, and native \emph{Julia} MLs for seamless integration of heterogeneous 
+libraries to create complex ensembles for robust time series prediction, clustering, and classification.
+\textbf{TSML} has the following features:
 \begin{enumerate}
 \item data type clustering/classification for automatic data discovery
 \item aggregation based on date/time interval
 \item imputation based on symmetric Nearest Neighbors
 \item statistical metrics for data quality assessment and classification input features
-\item ML wrapper with more than 100+ libraries from caret, scikitlearn, and julia
+\item ML wrapper with more than 100+ libraries from caret, scikit-learn, and julia
 \item date/value matrix conversion of 1-D time series using sliding windows to generate features for ML prediction
 \item pipeline API for high-level description of the processing workflow
 \item specific cleaning/normalization workflow based on data type
@@ -41,16 +44,12 @@ TSML has the following features:
 \item automatic segmentation of time-series data into matrix form for ML training and prediction 
 \item extensible architecture using just two main interfaces: fit and transform
 \item meta-ensembles for automatic feature and model selection
-\item support for distributed computation for scalability and speed
+\item support for distributed/threaded computation for scalability and speed
 \end{enumerate}
 
-The \textbf{TSML} package assumes a two-column input composed of \emph{dates} and \emph{values}. 
-The first part of the workflow aggregates values based on the specified date/time 
-interval which minimizes occurrence of missing values and noise. The aggregated 
-data is then left-joined to the complete sequence of dates in a specified date/time interval. 
-Remaining missing values are replaced by \textit{k} nearest neighbors where \textit{k} is the symmetric 
-distance from the location of missing value. This approach can be called several 
-times until there are no more missing values.
[…]
