+++ paper/ref.bib
@@ -58,3 +58,23 @@
   howpublished = {\url{https://github.com/ppalmes/CombineML.jl}},
   commit = {a94fd31f15df1a7eff1e1fdc2a067c67666c3654}
 }
+
+@misc{tsml2019,
+  author = {Palmes, P.},
+  title = {{TSML}: Time Series Machine Learning},
+  year = {2019},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/IBM/TSML.jl}}
+}
+
+@misc{decisiontree2008,
+  author = {Sadeghi, B and Coelho, L.P. },
+  title = {DecisionTree: Julia implementation of Decision Tree and Random Forest algorithms},
+  year = {2019},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/bensadeghi/DecisionTree.jl}}
+}
+
+
