### 2019-09-12 aefbe86
more discussions and public datasets analysis
