+++ paper/paper.tex
@@ -47,7 +47,7 @@ libraries to create complex ensembles for robust time series prediction, cluster
 \item support for distributed/threaded computation for scalability and speed
 \end{enumerate}
 
-The \textbf{TSML} package assumes a two-column input for any time series data composed of \emph{dates} and \emph{values}. The first part of the workflow aggregates values based on the specified date/time interval which minimizes occurrence of missing values and noise. The aggregated data is then left-joined to the complete sequence of dates in a specified date/time interval. Remaining missing values are replaced by \textit{k}-nearest neighbors (KNN) where \textit{k} is the symmetric distance from the location of missing value. This approach can be called several times until there are no more missing values.
+The \textbf{TSML} package assumes a two-column input for any time series data composed of \emph{dates} and \emph{values}. The first part of the workflow aggregates values based on the specified date/time interval which minimizes occurrence of missing values and noise. The aggregated data is then left-joined to the complete sequence of dates in a specified date/time interval. Remaining missing values are replaced by the median/mean or user-defined aggregation function of the \textit{k}-nearest neighbors (KNN) where \textit{k} is the symmetric distance from the location of missing value. This approach can be called several times until there are no more missing values.
 
 \vskip 6pt
 
@@ -72,7 +72,7 @@ There are more than hundred classifiers and regression functions available using
 
 \vskip 6pt
 To illustrate, below describes the main steps in using \textbf{TSML}.
-First, we create filters for csv reading, aggregation, imputation, data quality
+First, we create filters for csv reading, aggregation, imputation, and data quality
 assessment.
 
 \begin{lstlisting}[language = Julia]
@@ -146,7 +146,7 @@ of missing blocks count is now empty.
 
 \vskip 6pt
 
-We can also visualise our time series data using the \emph{plotter} filter instead of the \emph{statifier} as shown in Fig~\ref{fig:mplot}. Looking closely, you will see discontinuities in the plot due to blocks of missing data.
+We can also visualise our time series data using the \emph{Plotter} filter instead of the \emph{Statifier} as shown in Fig~\ref{fig:mplot}. Looking closely, you will see discontinuities in the plot due to blocks of missing data.
 
 \begin{lstlisting}[language = Julia]
 pltr=Plotter(Dict(:interactive => true))
@@ -163,7 +163,7 @@ transform!(plpipeline)
    \label{fig:mplot}
 \end{figure}
 
-Using the imputation pipeline described before, we can visualise the result of imputation by replacing the \emph{Statifier} with the \emph{Plotter} filter. Figure~\ref{fig:amplot} shows the plot after imputation which gets rid of missing data.
+Using the imputation pipeline described before, we can visualise the result by replacing the \emph{Statifier} with the \emph{Plotter} filter. Figure~\ref{fig:amplot} shows the plot after imputation which gets rid of missing data.
 
 \begin{lstlisting}[language = Julia]
 bplpipeline = Pipeline(Dict(
@@ -271,7 +271,7 @@ As seen in the plot, the values in daily monotonic time series is monotonically 
 
 \vskip 6pt
 
-Let us now reuse the \emph{Monotonicer} filter we used in the previous subsection to normalise the data.
+Let us now reuse the \emph{Monotonicer} filter in the previous subsection to normalise the data.
 
 \begin{lstlisting}[language = Julia]
 dailymonopipeline = Pipeline(Dict(
@@ -308,13 +308,13 @@ transform!(dailymonopipeline)
 
 \section{Time Series Classification}
 
-We can now use the knowledge we learned in setting up the \textbf{TSML}
+We can use the knowledge we learned in setting up the \textbf{TSML}
 \emph{pipeline} and \emph{filters} or \emph{transformers} to build higher level
[…]
