+++ paper/ref.bib
@@ -77,4 +77,12 @@
   howpublished = {\url{https://github.com/bensadeghi/DecisionTree.jl}}
 }
 
+@misc{tsmlextra2019,
+  author = {Palmes, P.},
+  title = {{TSMLextra}: External Machine Learning Libs for TSML},
+  year = {2019},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/ppalmes/TSMLextra.jl}}
+}
 
