+++ paper/paper.tex
@@ -708,15 +708,13 @@ One thing to note is that the Refrigeration dataset has almost the same number o
 \begin{table}
 \tbl{Electric Devices}{
 \begin{tabular}{|r|c|c|c|}\hline
-\textbf{Model} & \textbf{Mean} & \textbf{Std} & \textbf{Trials}\\\hline
+\textbf{Model} & \textbf{MeanFscore} & \textbf{Std} & \textbf{Trials}\\\hline
 c-rf & 0.56 & 0.00 & 10\\\hline
 c-treebag & 0.54 & 0.01 & 10\\\hline
 s-extratree & 0.52 & 0.01 & 10\\\hline
 j-vote & 0.51 & 0.01 & 10 \\\hline
 j-rf & 0.51 & 0.01 & 10 \\\hline
 ... & ... & ... & ...\\\hline
-j-ada & 0.22 & 0.00 & 10 \\\hline
-c-rpart & 0.06 & 0.00 & 10 \\\hline
 \end{tabular}}
 \label{tab:elec}
 \end{table}
@@ -724,15 +722,13 @@ c-rpart & 0.06 & 0.00 & 10 \\\hline
 \begin{table}
 \tbl{Refrigeration Devices}{
 \begin{tabular}{|r|c|c|c|}\hline
-\textbf{Model} & \textbf{Mean} & \textbf{Std} & \textbf{Trials}\\\hline
+\textbf{Model} & \textbf{MeanFscore} & \textbf{Std} & \textbf{Trials}\\\hline
 s-gb & 0.51 & 0.00 & 10\\\hline
 c-rf & 0.51 & 0.01 & 10\\\hline
 c-treebag & 0.50 & 0.01 & 10\\\hline
 j-super & 0.50 & 0.02 & 10 \\\hline
 s-rf & 0.48 & 0.03 & 10 \\\hline
 ... & ... & ... & ...\\\hline
-j-ada & 0.33 & 0.00 & 10 \\\hline
-c-rpart & 0.12 & 0.00 & 10 \\\hline
 \end{tabular}}
 \label{tab:ref}
 \end{table}
@@ -740,15 +736,13 @@ c-rpart & 0.12 & 0.00 & 10 \\\hline
 \begin{table}
 \tbl{Earthquakes}{
 \begin{tabular}{|r|c|c|c|}\hline
-\textbf{Model} & \textbf{Mean} & \textbf{Std} & \textbf{Trials}\\\hline
+\textbf{Model} & \textbf{MeanFscore} & \textbf{Std} & \textbf{Trials}\\\hline
 c-treebag & 0.86 & 0.01 & 10\\\hline
 c-rf & 0.86 & 0.00 & 10\\\hline
 c-rpart & 0.86 & 0.00 & 10\\\hline
 s-extratree & 0.85 & 0.01 & 10 \\\hline
 s-gb & 0.85 & 0.00 & 10 \\\hline
 ... & ... & ... & ...\\\hline
-c-svmlinear & 0.77 & 0.00 & 10 \\\hline
-j-pt & 0.76 & 0.01 & 10 \\\hline
 \end{tabular}}
 \label{tab:earth}
 \end{table}
@@ -756,20 +750,18 @@ j-pt & 0.76 & 0.01 & 10 \\\hline
 \begin{table}
 \tbl{FordB}{
 \begin{tabular}{|r|c|c|c|}\hline
-\textbf{Model} & \textbf{Mean} & \textbf{Std} & \textbf{Trials}\\\hline
+\textbf{Model} & \textbf{MeanFscore} & \textbf{Std} & \textbf{Trials}\\\hline
[…]
