+++ paper/ref.bib
@@ -86,3 +86,19 @@
   howpublished = {\url{https://github.com/ppalmes/TSMLextra.jl}}
 }
 
+@Article{Breiman2001,
+author="Breiman, Leo",
+title="Random Forests",
+journal="Machine Learning",
+year="2001",
+month="Oct",
+day="01",
+volume="45",
+number="1",
+pages="5--32",
+abstract="Random forests are a combination of tree predictors such that each tree depends on the values of a random vector sampled independently and with the same distribution for all trees in the forest. The generalization error for forests converges a.s. to a limit as the number of trees in the forest becomes large. The generalization error of a forest of tree classifiers depends on the strength of the individual trees in the forest and the correlation between them. Using a random selection of features to split each node yields error rates that compare favorably to Adaboost (Y. Freund {\&} R. Schapire, Machine Learning: Proceedings of the Thirteenth International conference, ***, 148--156), but are more robust with respect to noise. Internal estimates monitor error, strength, and correlation and these are used to show the response to increasing the number of features used in the splitting. Internal estimates are also used to measure variable importance. These ideas are also applicable to regression.",
+issn="1573-0565",
+doi="10.1023/A:1010933404324",
+url="https://doi.org/10.1023/A:1010933404324"
+}
+
