+++ paper/paper.tex
@@ -782,11 +782,11 @@ Comparing the performances of different classifiers among the four problems indi
 
 \vskip 6pt
 
-Among the classifiers, the \emph{Random Forest} from \emph{Caret} is the top performer in all problems. On the other hand, \emph{TreeBag} from \emph{Caret} and \emph{Gradient Boosting} from \emph{ScikitLearn} are the top performers in Refrigeration Devices and Earthquakes, respectively. It is interesting to note that among the \emph{Random Forest} implementations, the top performer \emph{Caret's} version was written by Breiman\cite{Breiman2001} who was the original author of the said algorithm in \emph{fortran}. The superior performance of Breiman's Random Forest can be highlighted in the FordB classification performance. The next best which is \emph{k}NN is 4\% lower. In all other cases, the next best performer has almost same performance with  Breiman's \emph{Random Forest}.
+Among the classifiers, the \emph{Random Forest} from \emph{Caret} is the top performer in all problems. On the other hand, \emph{TreeBag} from \emph{Caret} and \emph{Gradient Boosting} from \emph{ScikitLearn} are the top performers in Refrigeration Devices and Earthquakes, respectively. It is interesting to note that among the \emph{Random Forest} implementations, the top performer \emph{Caret's} version was written by Breiman\cite{Breiman2001} who was the original author of the said algorithm in \emph{fortran}. The superior performance of Breiman's \emph{Random Forest} can be highlighted in the FordB classification performance.  \emph{k}NN, the next best classifier is 4\% lower than Breiman's \emph{Random Forest}. In all other cases, the next best performer has almost same performance with  Breiman's \emph{Random Forest}.
  
 \vskip 6pt
  
-As we expected based on past studies, the different ensemble models dominated the top 5 performers. While not shown in the table, the most consistent worst performer is the \emph{Caret's} \emph{RPart}. This is consistent to Breiman's observation that the Random Forest performs well by using unstable or inferior ML models to its leaves. In \emph{Caret's} \emph{Random Forest}, its leaves are composed of \emph{RPart} models which has the poorest performance in all problems. The combination of boosting and bagging weak learners such \emph{RPart} makes the \emph{Random Forest} robust from datasets imbalances.
+As we expected based on past studies, the different ensemble models dominated the top 5 performers. While not shown in the table, the most consistent worst performer is the \emph{Caret's} \emph{RPart}. This is consistent to Breiman's observation that \emph{Random Forest} performs well by using unstable or inferior ML models in its leaves. In \emph{Caret's} \emph{Random Forest}, the leaves are composed of \emph{RPart} models which has the poorest performance in all problems. The combination of boosting and bagging weak learners such as \emph{RPart} make the \emph{Random Forest} robust from datasets imbalances. 
  
 \section{Summary and Conclusion}
 Packages for time series analysis are becoming important tools with the rapid proliferation of sensor data brought about by \textbf{IoT}.  We created \textbf{TSML} as a time series machine learning framework which can easily be extended to handle large volume of time series data. 
