### 2019-09-12 fed8430
minor edits
