+++ paper/paper.tex
@@ -47,7 +47,7 @@ libraries to create complex ensembles for robust time series prediction, cluster
 \item support for distributed/threaded computation for scalability and speed
 \end{enumerate}
 
-The \textbf{TSML} package assumes a two-column input for any time series data composed of \emph{dates} and \emph{values}. The first part of the workflow aggregates values based on the specified date/time interval which minimizes occurrence of missing values and noise. The aggregated data is then left-joined to the complete sequence of dates in a specified date/time interval. Remaining missing values are replaced by the median/mean or user-defined aggregation function of the \textit{k}-nearest neighbors (KNN) where \textit{k} is the symmetric distance from the location of missing value. This approach can be called several times until there are no more missing values.
+The \textbf{TSML} package assumes a two-column input for any time series data composed of \emph{dates} and \emph{values}. The first part of the workflow aggregates values based on the specified date/time interval which minimizes occurrence of missing values and noise. The aggregated data is then left-joined to the complete sequence of dates in a specified date/time interval. Remaining missing values are replaced by the median/mean or user-defined aggregation function of the \textit{k}-nearest neighbors (\emph{KNN}) where \textit{k} is the symmetric distance from the location of missing value. This approach can be called several times until there are no more missing values.
 
 \vskip 6pt
 
@@ -323,8 +323,7 @@ Time series classification is a significant first step for optimal prediction an
 Identifying the correct sensor data types can help in the choice of what the most optimal prediction 
 model to use for actuation or pre-emption to minimise wastage in resource utilisation.
 To successfully perform the latter operations, it is necessary to identify first the time series
-type so that appropriate model and cleaning routines can be selected for optimal model performance . 
-The  \emph{TSClassifier} filter aims to address this problem and its usage is described below.
+type so that appropriate model and cleaning routines can be selected for optimal model performance. The  \emph{TSClassifier} filter aims to address this problem and its usage is described below.
 
 \vskip 6pt
 
@@ -556,18 +555,18 @@ The data used in the experiment are sample snapshots of the data in our building
 
 \vskip 6pt
 
-Figures~\ref{fig:sim}~and~\ref{fig:performance}  show a snapshot of running workers exploiting the distributed library of Julia and the classification performance of each model, respectively. There are 8 workers running in parallel over 12 different machine learning classifiers. 
+Figures~\ref{fig:sim}~and~\ref{fig:performance}  show a snapshot of running workers exploiting the distributed library of \emph{Julia} and the classification performance of each model, respectively. There are 8 workers running in parallel over 12 different machine learning classifiers. 
 
 \vskip 6pt
 
 From the results, \emph{ExtraTree} from \emph{Scikit-Learn} has the best performance with 91.67\% accuracy followed by \emph{TreeBag} and \emph{SVMLinear} from \emph{Caret} library with 83.33 \% accuracy for both. With this workflow, it becomes trivial to search for optimal model by running them in parallel relying on \emph{Julia} to do the low-level tasks of scheduling and queueing as well as making sure that  the dynamically available compute resources such as cpu cores and memory resources are fairly optimised.           
 
 \subsection{Parallel TSML Using Threads Workflow}
-With \emph{Julia} 1.3, lightweight multi-threading support in Julia becomes possible. We will be using the pure \emph{Julia}-written ML models because installing external dependencies such as \emph{Caret} MLs through \emph{RCall} package has some issues with the alpha version of \emph{Julia} 1.3 at this point in time. We will update this documentation and add more MLs once the issues are resolved.
[…]
