### 2019-09-16 e6672d8
remove travis build for this branch
