+++ .travis.yml
@@ -1,45 +1,0 @@
-# Documentation: http://docs.travis-ci.com/user/languages/julia/
-language: julia
-os:
-  - linux
-  - osx
-julia:
-  - 1.0
-  - 1.1
-  - nightly
-env:
-  - JULIA_NUM_THREADS=8
-matrix:
-  fast_finish: true
-  allow_failures:
-    - julia: nightly
-
-notifications:
-  email: false
-
-after_success:
-  - if [[ $TRAVIS_OS_NAME = linux ]]; then
-      julia --project=coverage/ -e 'using Pkg; Pkg.instantiate();
-          using Coverage; Codecov.submit(Codecov.process_folder())';
-    fi
-
-jobs:
-  include:
-    - stage: "HTML Documentation"
-      julia: 1.1
-      os: osx
-      script:
-        - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
-        - julia --project=docs/ docs/make.jl
-      name: "HTML"
-      after_success: skip
-      services: docker
-   # - stage: "PDF Documentation"
-   #   julia: 1.1
-   #   os: osx
-   #   script:
-   #     - julia --project=docs/pdf/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
-   #     - julia --project=docs/pdf/ docs/pdf/make.jl
-   #   name: "PDF"
-   #   after_success: skip
-   #   services: docker
