### 2019-09-19 #51 @matbesancon
@theogf @christianpeel you can continue the review here
