### 2019-09-19 IBM/TSML.jl#88 @christianpeel
DOIs for references
My understanding of the [review](https://github.com/JuliaCon/proceedings-review/issues/51) instructions for the JuliaCon proceedings is that your references bib file should contain DOI identifiers, e.g. for the Julia paper by Bezanson et al.
