### 2019-09-19 IBM/TSML.jl#89 @christianpeel
Colloquial names are not so meaningful
Calvin and Hobbes certainly were able to [invent](https://calvinandhobbes.fandom.com/wiki/Transmogrifier) a -ier word, yet this is not a comic strip, and I found your colloquial verbs to be more cutely distracting than informative. Specifically Datefier, DateValizer, Matrifier, Outliernicer, Statifier, and Monotonicer seem to describe objects, rather than functions.

I saw these as part of the JuliaCon proceedings [review](https://github.com/JuliaCon/proceedings-review/issues/51), yet there's no need to change anything for that review. This comment is just a minor comment that you're welcome to ignore.
