### 2019-09-19 #51 @christianpeel
I added a minor issue ([DOIs](https://github.com/IBM/TSML.jl/issues/88)) to the TSML repo.
