### 2019-09-24 IBM/TSML.jl#89 @christianpeel
Thanks for the comments.  There's no need to change anything.
