### 2019-09-24 #51 @ppalmes
hi @matbesancon, in reference to DOI issue raised by @christianpeel , i added \usepackage{doi} but it seems that the bibstyle does not include the url and doi info in the bibliography although ref.bin has doi entries. Is there a command or options in the latex i need to add to include doi in the bibliography entry of final pdf?
