### 2019-09-24 8b5edb2
add more doi entries
