+++ paper/paper.tex
@@ -3,6 +3,8 @@
 \documentclass{juliacon}
 \setcounter{page}{1}
 
+\usepackage{doi}
+
 \hyphenation{TSMLextra}
 \hyphenation{TSML}
 
