+++ paper/ref.bib
@@ -9,32 +9,36 @@
 
 
 @article{bezanson2017julia,
-	Author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
-	Journal = {SIAM review},
-	Number = {1},
-	Pages = {65--98},
-	Publisher = {SIAM},
-	Title = {Julia: A fresh approach to numerical computing},
-	Volume = {59},
-	Year = {2017}}
+  Author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
+  Journal = {SIAM review},
+  Number = {1},
+  Pages = {65--98},
+  Publisher = {SIAM},
+  Title = {Julia: A fresh approach to numerical computing},
+  url = {https://doi.org/10.1137/141000671},
+  doi = {10.1137/141000671},
+  Volume = {59},
+  Year = {2017}
+}
 
 @article{Moore1990,
-	Author = {Moore, James T. and Bard, Jonathan F.},
-	Doi = {10.1287/opre.38.5.911},
-	Issn = {0030-364X},
-	Journal = {Operations Research},
-	Keywords = {games: noncooperative,integer: branch-and-bound algorithms,programming},
-	Mendeley-Groups = {bilevel - hierarchical},
-	Month = {oct},
-	Number = {5},
-	Pages = {911--921},
-	Publisher = {INFORMS},
-	Title = {{The Mixed Integer Linear Bilevel Programming Problem}},
-	Url = {http://pubsonline.informs.org/doi/abs/10.1287/opre.38.5.911},
-	Volume = {38},
-	Year = {1990},
-	Bdsk-Url-1 = {http://pubsonline.informs.org/doi/abs/10.1287/opre.38.5.911},
-	Bdsk-Url-2 = {https://doi.org/10.1287/opre.38.5.911}}
+  Author = {Moore, James T. and Bard, Jonathan F.},
+  Doi = {10.1287/opre.38.5.911},
+  Issn = {0030-364X},
+  Journal = {Operations Research},
+  Keywords = {games: noncooperative,integer: branch-and-bound algorithms,programming},
+  Mendeley-Groups = {bilevel - hierarchical},
+  Month = {oct},
+  Number = {5},
+  Pages = {911--921},
+  Publisher = {INFORMS},
+  Title = {{The Mixed Integer Linear Bilevel Programming Problem}},
+  Url = {http://pubsonline.informs.org/doi/abs/10.1287/opre.38.5.911},
+  Volume = {38},
+  Year = {1990},
+  Bdsk-Url-1 = {http://pubsonline.informs.org/doi/abs/10.1287/opre.38.5.911},
+  Bdsk-Url-2 = {https://doi.org/10.1287/opre.38.5.911}
+}
 
 @misc{orchestra2014,
   author = {Jenkins, S.},
@@ -45,7 +49,6 @@
   howpublished = {\url{https://github.com/svs14/Orchestra.jl}},
   commit = {1d01e4b94e08fa19ec0ce41f0d0a1bdc4516bd4f}
 }
-
 
 @misc{combineml2016,
   author = {Palmes, P.},
@@ -87,18 +90,18 @@
 }
 
 @Article{Breiman2001,
-author="Breiman, Leo",
-title="Random Forests",
-journal="Machine Learning",
-year="2001",
-month="Oct",
-day="01",
-volume="45",
-number="1",
-pages="5--32",
[…]
