### 2019-09-24 #51 @Nosferican
See https://github.com/JuliaCon/JuliaConSubmission.jl/issues/7#issuecomment-496266177.
