### 2019-09-24 41b8cba
remove doi package
