+++ paper/paper.tex
@@ -3,8 +3,6 @@
 \documentclass{juliacon}
 \setcounter{page}{1}
 
-\usepackage{doi}
-
 \hyphenation{TSMLextra}
 \hyphenation{TSML}
 
