+++ paper/ref.bib
@@ -42,7 +42,7 @@
 
 @misc{orchestra2014,
   author = {Jenkins, S.},
-  title = {Orchestra.jl},
+  title = {Orchestra: Heterogeneous ensemble learning for Julia},
   year = {2014},
   publisher = {GitHub},
   journal = {GitHub repository},
