### 2019-09-24 #51 @ppalmes
@christianpeel, i added the doi entries but bibtex does not print this out in the final bib entry of pdf as noted in the issue https://github.com/JuliaCon/proceedings-review/issues/51#issuecomment-534626356
