### 2019-09-24 #51 @christianpeel
Thanks for adding the DOIs, which is enough for me. I marked the requirement as satisfied; fixing the .bib format should not be your problem to solve. The same thing came up when I was reviewing Kronecker.jl, and the issue you link shows that others have the problem as well.

I suggest that @matbesancon remove this DOI requirement from all current and future JuliaCon review submissions until a way to satisfy it is provided.
