### 2019-09-24 IBM/TSML.jl#90 @christianpeel
Context and Results for JuliaCon Proceedings
In the paper you give several examples of the use of TSML, but I didn't catch what we should be comparing TSML against.  Specifically, you have results that compare different tools that can be accessed from within TSML, but no results from some other Julia or non-Julia packages to show that TSML is better or worse in any way.  It's certainly possible that I missed it; could you help me out?

I.e. I'm looking at the "Context" and "Results" items in the JuliaCon [Proceedings issue](https://github.com/JuliaCon/proceedings-review/issues/51).
