### 2019-09-24 IBM/TSML.jl#88 @christianpeel
Adding the DOIs to the .bib file is enough.
