### 2019-09-24 IBM/TSML.jl#90 @ppalmes
Hi @christianpeel,

In the latest pdf:

https://github.com/JuliaCon/proceedings-papers/blob/jcon.00051/jcon.00051/10.21105.jcon.00051.pdf

Section 5 compares the different MLs from R, Julia, and ScikitLearn. TSML uses a wrapper for R caret and Python ScikitLearn so that they can easily be used without much difference in julia syntax by calling the fit! and transform! multidispatch functions. The comparison is more of showing the feasibility of combining different MLs from different libraries.
