### 2019-09-25 IBM/TSML.jl#90 @christianpeel
I apologize for my confusion.   To be clear, is there a package like TSML.jl in R or Python or Swift that has similar functionality that you might compare against?

My understanding is that Section 5 of your paper compares the performance of specific tools on specific models; to be clear, this is useful. I may have missed it, but I do not recall seeing a comparison of TSML.jl with any similar packages.   If indeed there is some other packages comparable to TSML.jl in Julia or R or Python or some other language, then I'd like to know how TSML.jl compares against them. If TSML.jl is unique, then that is useful info and could go in your paper.
