### 2019-09-25 IBM/TSML.jl#90 @ppalmes
TSML.jl is like scikitlearn in python and caret in R. all of them create a common wrapper for MLs in their respective languages. what is unique with TSML is that it has a wrapper to use the MLs in both caret and scikitlearn. In the comparison, you have performance of MLs from R thru caret and scikitlearn from Python against ensembles written in Julia. the advantage of TSML is that you can run libraries from R and python and combine them for complex ensembles. so far, i have not seen a metapackage that combines all MLs written in R, Python, and Julia working together seamlessly. Since each ML library in R are compiled, calling them from R and from Julia thru RCall are not much different. The comparisons can be considered among different MLs.
