### 2019-09-25 IBM/TSML.jl#90 @ppalmes
For example, these are the possible ML models in R which can be called using the caret common wrapper function of train and predict: https://topepo.github.io/caret/available-models.html

Although I have not exhaustively test all MLs in R, TSML also can call them using the fit! and transform! multi-dispatch functions of TSML using RCall package as the middleware between Julia and R communication.
