### 2019-09-25 IBM/TSML.jl#90 @christianpeel
Perhaps it would pay to look at other metapackages like JuMP.jl, Plots.jl, DifferentialEquations.jl to see what words they use to describe themselves.
