### 2019-09-25 IBM/TSML.jl#90 @ppalmes
The ultimate goal of the ML part in TSML is to automatically pick the right ensemble architecture where the elements are coming from the different ML libraries. The challenge is to have a metaML optimize both the parameters and architecture of  the bunch of MLs for a given problem. This is similar to autoML but the focus is more on automatic meta modeling instead entire pipeline which the autoML targets. This is in contrast to the Plots.jl or DifferentialEquations.jl which does not involve automatic modeling and learning.
