### 2019-09-25 IBM/TSML.jl#90 @christianpeel
I mentioned JuMP.jl, Plots.jl, and DifferentialEquations.jl because they are metapackages. I did not intend any direct comparison of the functionality of the packages with TSML.jl. I only mentioned them as a possible place to look for words that describe a metapackage; i.e. words other than "package".
