### 2019-09-25 IBM/TSML.jl#90 @ppalmes
yes, i know what you mean. i think ensemble is more appropriate than meta-package because the combination is data-driven and dynamic. thanks for the suggestions. i’ll try to improve the explanation. the idea is an evolution of different versions. i just started with mostly filters and now extending it to ensemble learning.
