### 2019-09-30 #51 @ppalmes
@christianpeel, @theogf, @matbesancon : is the paper still under review? thanks!
