### 2019-10-09 IBM/TSML.jl#91 @theogf
[REVIEW] JuliaCon Paper
If it's okay I am going to gather all my comments for the review in https://github.com/JuliaCon/proceedings-review/issues/51  here:

First it's a very nice package which is for sure extremely useful. I really like the pipeline composition style and how easy and flexible it is to construct it. The output part is also well made and give a clear understanding of the results.
But I still have a few comments about the paper :

 - [x] The structure `fit!` and `transform!` does not make always sense to me, I think it would be good to clarify better what they really mean, and eventually list the differences for the different types of "modules", i.e. if it is a reader, a filter or a classifier etc... It would also help for people wanting to create new filters. You also never explain what should be the output/input of each module in the pipeline.

- [x] I would like to repeat @christianpeel point in #90: you give extremely nice examples on how to use package on real data, but giving an analysis on the individual results of each algorithm is not really necessary. I also made a small research and it seems that there exists  no (open-source) tool comparable to yours :+1: !

More on the API side/documentation :

- [x] I also share @christianpeel 's opinion on the naming of the filters, for an external person it is not so clear what they do from first sight, but as in #89 this is a minor comment.

- [x] Connected to the last point, is there a reason you always pass `Dict` to your modules? When using autocomplete tools, in Juno for instance, it is always more practical to see what arguments are possible, especially when one is not familiar with all the packages.

- [x] Similarly it would be nice to have inline documentation: for example calling `? DataValizer` would give a direct explanation of the filter and the possible arguments.

PS : Is there any reason why you don't give the option to simply drop the missing values?
