### 2019-10-14 IBM/TSML.jl#91 @ppalmes
Hi @theogf,

Thanks for your nice comments and suggestions.

Regarding fit! and transform!, the type hierarchy looks like this:
```Any Machine Learner <: TSLearner <: Transformer```
```Any Filter <: Transformer```
```Pipeline <: Transformer```

`fit!` in machine learner requires training of parameters for regression or classification. `fit!` in machine learner requires both input and output. On the other hand, fit! in a filter process only the input because it doesn't learn any mapping between input and output. fit! in a filter is typically a straightforward computation of some statistics such as normalization stats, range, or PCA/ICA parameters for space embedding, etc.

The `transform!` in both machine learner and filter apply these computed or learned parameters for the new dataset. In many types of filter, fit! just merely checks for errors or initialize some variables to be used but all important operations are done during the `transform!`. In the machine learner, most critical operations are done during `fit!` to learn parameters and `transform!` is just application of parameters to new data.
