### 2019-10-16 #51 @matbesancon
@ppalmes did you complete and adapt the paper to the review from @theogf ?
