### 2019-10-16 #51 @ppalmes
hi @matbesancon, i have addressed it here: https://github.com/IBM/TSML.jl/issues/91

Most comments and suggestions are more of clarifications.

From the comments and suggestions, the inline documentation is scheduled in the next release but should not affect the paper coverage here.
