### 2019-10-18 658f1d6
separate transfomers to basefilters and valdatefilters
