### 2019-10-18 a5b3b2d
inline docs of valdate filters
