+++ src/valdatefilters.jl
@@ -22,16 +22,43 @@ export BzCSVDateValReader
 
 
 
-# Convert a 1-D timeseries into sliding window matrix for ML training
-# using Plots
 
 const gAggDict = Dict(
     :median => Statistics.median,
     :mean =>   Statistics.mean,
     :maximum => Statistics.maximum,
-    :minimum => Statistics.minimum
+    :minimum => Statistics.minimum,
+    :sum => sum
 )
 
+"""
+    Matrifier(Dict(
+       Dict(
+        :ahead => 1,
+        :size => 7,
+        :stride => 1,
+      )
+    )
+
+Converts a 1-D timeseries into sliding window matrix for ML training:
+- `:ahead` => steps ahead to predict
+- `:size` => size of sliding window
+- `:stride` => amount of overlap in sliding window
+
+Example:
+
+    mtr = Matrifier(Dict(:ahead=>24,:size=>24,:stride=>5))
+    lower = DateTime(2017,1,1)
+    upper = DateTime(2017,1,5)
+    dat=lower:Dates.Hour(1):upper |> collect
+    vals = 1:length(dat)
+    x = DataFrame(Date=dat,Value=vals)
+    fit!(mtr,x)
+    res = transform!(mtr,x)
+
+
+Implements: `fit!`, `transform`
+"""
 mutable struct Matrifier <: Transformer
   model
   args
@@ -46,6 +73,12 @@ mutable struct Matrifier <: Transformer
   end
 end
 
+
+"""
+    fit!(mtr::Matrifier,xx::T,y::Vector=Vector()) where {T<:Union{Matrix,Vector,DataFrame}}
+
+Checks and validate inputs are in correct structure
+"""
 function fit!(mtr::Matrifier,xx::T,y::Vector=Vector()) where {T<:Union{Matrix,Vector,DataFrame}}
   typeof(xx) <: DataFrame || error("input is not a dataframe")
   x = deepcopy(xx.Value)
@@ -53,6 +86,11 @@ function fit!(mtr::Matrifier,xx::T,y::Vector=Vector()) where {T<:Union{Matrix,Ve
   mtr.model = mtr.args
 end
 
+"""
+    transform!(mtr::Matrifier,xx::T) where {T<:Union{Matrix,Vector,DataFrame}}
+
+Applies the parameters of sliding windows to create the corresponding matrix
+"""
 function transform!(mtr::Matrifier,xx::T) where {T<:Union{Matrix,Vector,DataFrame}}
   typeof(xx) <: DataFrame || error("input is not a dataframe")
   x = deepcopy(xx.Value)
@@ -84,7 +122,31 @@ end
 
 ### ====
 
-# Convert a 1-D date series into sliding window matrix for ML training
+
+"""
+    Dateifier(args=Dict())
+       Dict(
+        :ahead => 1,
+        :size => 7,
+        :stride => 1
+       )
+    )
+
+Converts a 1-D date series into sliding window matrix for ML training
+
+Example: 
+
+    dtr = Dateifier(Dict())
+    lower = DateTime(2017,1,1)
+    upper = DateTime(2018,1,31)
+    dat=lower:Dates.Day(1):upper |> collect
+    vals = rand(length(dat))
+    x=DataFrame(Date=dat,Value=vals)
+    fit!(dtr,x)
+    res = transform!(dtr,x)
+
+Implements: `'fit!`, `transform!`
+"""
 mutable struct Dateifier <: Transformer
   model
   args
@@ -99,6 +161,11 @@ mutable struct Dateifier <: Transformer
   end
 end
 
+"""
+    fit!(dtr::Dateifier,xx::T,y::Vector=[]) where {T<:Union{Matrix,Vector,DataFrame}}
+
[…]
