+++ docs/Project.toml
@@ -4,6 +4,7 @@ DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 DocumenterTools = "35a29f4d-8980-5a13-9543-d66fff28ecb8"
+IJulia = "7073ff75-c697-5162-941a-fcdaad2a7d2a"
 Latexify = "23fbe1c1-3f47-55db-b15f-69d7ec21a316"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 TSML = "198dc43e-9e51-5cd7-9d40-d9794d335912"
