### 2019-10-18 0f43398
inline docs refinement
