+++ src/baseline.jl
@@ -13,6 +13,17 @@ export Baseline,Identity
 
 using StatsBase: mode
 
+
+"""
+    Baseline(
+       default_args = Dict(
+          :output => :class,
+          :strat => mode
+       )
+    )
+
+Baseline model that returns the mode during classification.
+"""
 mutable struct Baseline <: TSLearner
     model
     args
@@ -26,14 +37,29 @@ mutable struct Baseline <: TSLearner
     end
 end
 
+"""
+    fit!(bsl::Baseline,x::Matrix,y::Vector)
+
+Get the mode of the training data.
+"""
 function fit!(bsl::Baseline,x::Matrix,y::Vector)
     bsl.model = bsl.args[:strat](y)
 end
 
+"""
+    transform!(bsl::Baseline,x::Matrix)
+
+Return the mode in classification.
+"""
 function transform!(bsl::Baseline,x::Matrix)
     fill(bsl.model,size(x,1))
 end
 
+"""
+    Identity(args=Dict())
+
+Returns the input as output.
+"""
 mutable struct Identity <: Transformer
     model
     args
@@ -44,10 +70,20 @@ mutable struct Identity <: Transformer
     end
 end
 
+"""
+    fit!(idy::Identity,x::Matrix,y::Vector)
+    
+Does nothing.
+"""
 function fit!(idy::Identity,x::Matrix,y::Vector)
     nothing
 end
 
+"""
+    transform!(idy::Identity,x::Matrix)
+
+Return the input as output.
+"""
 function transform!(idy::Identity,x::Matrix)
     return x
 end

[…]
