### 2019-10-18 33634e7
disable fast finish
