+++ .travis.yml
@@ -10,7 +10,7 @@ julia:
 env:
   - JULIA_NUM_THREADS=8
 matrix:
-  fast_finish: true
+  #fast_finish: true
   allow_failures:
     - julia: nightly
 
