+++ Project.toml
@@ -1,7 +1,7 @@
 name = "TSML"
 uuid = "198dc43e-9e51-5cd7-9d40-d9794d335912"
 authors = ["Paulito Palmes <ppalmes@gmail.com>"]
-version = "2.3.9"
+version = "2.4.0"
 
 [deps]
 ArgParse = "c7e460c6-2fb9-53a9-8c5b-16f535851c63"
@@ -28,6 +28,22 @@ StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 
 [compat]
 julia = "1"
+CSV = "0.4, 0.5"
+Interact = "0.9, 0.10"
+GR = "0.35, 0.36, 0.37, 0.38, 0.39, 0.40, 0.41"
+JSON2 = "0.2, 0.3"
+HTTP = "0.6, 0.7, 0.8"
+IterTools = "1.1, 1.2"
+MLDataUtils = "0.2, 0.3, 0.4, 0.5"
+StatsBase = "0.25, 0.26, 0.27, 0.28, 0.29, 0.30, 0.31, 0.32"
+DecisionTree = "0.8, 0.9"
+Plots = "0.25, 0.26, 0.27"
+CodecBzip2 = "0.4, 0.5, 0.6"
+FileIO = "1.0"
+DataFrames = "0.17, 0.18, 0.19"
+Reexport = "0.1, 0.2"
+MLBase = "0.6, 0.7, 0.8"
+ArgParse = "0.5, 0.6"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
