+++ Project.toml
@@ -29,21 +29,21 @@ StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 [compat]
 julia = "1"
 CSV = "0.4, 0.5"
-Interact = "≥ 0.9"
-GR = "≥ 0.35"
-JSON2 = "≥ 0.2"
-HTTP = "≥ 0.6"
-IterTools = "≥ 1.1"
-MLDataUtils = "≥ 0.2"
-StatsBase = "≥ 0.25"
-DecisionTree = "≥ 0.8"
-Plots = "≥ 0.25"
-CodecBzip2 = "≥ 0.4"
-FileIO = "≥ 1.0"
-DataFrames = "≥ 0.17"
-Reexport = "≥ 0.1"
-MLBase = "≥ 0.6"
-ArgParse = "≥ 0.5"
+Interact = "0.9, 0.10"
+GR = "0.35, 0.36, 0.37, 0.38, 0.39, 0.40, 0.41"
+JSON2 = "0.2, 0.3"
+HTTP = "0.6, 0.7, 0.8"
+IterTools = "1.1, 1.2"
+MLDataUtils = "0.2, 0.3, 0.4, 0.5"
+StatsBase = "0.25, 0.26, 0.27, 0.28, 0.29, 0.30, 0.31, 0.32"
+DecisionTree = "0.8, 0.9"
+Plots = "0.25, 0.26, 0.27"
+CodecBzip2 = "0.4, 0.5, 0.6"
+FileIO = "1.0"
+DataFrames = "0.17, 0.18, 0.19"
+Reexport = "0.1, 0.2"
+MLBase = "0.6, 0.7, 0.8"
+ArgParse = "0.5, 0.6"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
