### 2019-10-19 f7ed723
add julia 1.2 testing
