+++ .travis.yml
@@ -6,6 +6,7 @@ os:
 julia:
   - 1.0
   - 1.1
+  - 1.2
   - nightly
 env:
   - JULIA_NUM_THREADS=8
