### 2019-10-20 IBM/TSML.jl#91 @ppalmes
@theogf, @christianpeel, @matbesancon: The just released TSML 2.3.9 now has inline documentation for  the types and most important functions. It includes inline examples too.
