### 2019-10-21 5333658
add more explaination about transfomer, filter, learner, pipeline
