### 2019-10-21 #51 @ppalmes
Here's a snippet of the added description  in the paper to explain more about Transformers, Filters, Learners, Pipeline, fit and transform:

------------

All major data types in TSML is a subtype of Transformer. There are two major types of transformers, namely: filters for data processing and learners for machine learning. Both transformers implement fit! and transform! multidispatch functions.

Filters are normally used for pre-processing tasks such as: imputation, normalization, feature extraction, feature transformation, scaling, etc. Consequently, filter’s fit! and transform! functions expect one argument which represents an input data for feature processing. Each data type must implement fit! and transform! although in many cases of preprocessing data, only transform! operation is needed. The reason for this is explained below in the discussion of Pipeline implementation. All filters are direct subtypes of Transformer.

On the other hand, a learner requires learning of parameters between input and output. The training part to learn the parameters is handled by fit! function and transform! for the prediction. All learners are subtypes of TSLearner <: Transformer.

TSML workflow borrows the idea of the Unix pipeline[3, 4]. The main elements in a TSML pipeline are series of filters or transform- ers each performing one specific task and does it well. In machine learning tasks for prediction or classification, a series of filters will be terminated by a machine learner which learns the mapping be tween the extracted features done by the filters together with the output argument.

Pipeline is also a subtype of Transformer. It is a container of transformers to perform pre-processing, training, and prediction or classification. The fit! function in Pipeline iteratively calls the fit! and transform! functions in a series of filters with possibly a learner at the endpoint. If the last transformer is a learner, the last transformed output from a series of filters will become input features of the learner for training.

The pipeline expects two arguments: input and output. The input argument is processed by filters while the transformed input together with the output arguments will be processed by the learner to learn the input and output mapping. During the prediction task, transform! in the Pipeline iteratively calls the transform! operations in each filter and learner. The transform operation is direct application of the parameters computed during normalization, scaling, training, etc. to the new data. If the last element in the pipeline during transform is a learner, it performs prediction or classification. Otherwise, transform operation acts as a feature extractor/transformer if they are composed of filters only.
