+++ paper/paper.tex
@@ -27,7 +27,7 @@
 \section{Introduction}
 
 \textbf{TSML}\cite{tsml2019} is a \emph{Julia}\cite{bezanson2017julia} package for time series data processing, classification, and prediction. 
-It provides common API for ML libraries from Python's \emph{Scikit-Learn}, 
+It provides common API for ML (Machine Learning) libraries from Python's \emph{Scikit-Learn}, 
 R's \emph{Caret}, and native \emph{Julia} MLs for seamless integration of heterogeneous 
 libraries to create complex ensembles for robust time series prediction, clustering, and classification.
 \textbf{TSML} has the following features:
@@ -60,37 +60,38 @@ the size and stride of the sliding window. The final part joins
  
  \vskip 6pt
  
-\textbf{TSML} uses a pipeline which iteratively calls the \emph{fit} and \emph{transform}
+\textbf{TSML} uses a pipeline which iteratively calls the \texttt{fit!} and \texttt{transform!}
 families of functions relying on \emph{multiple dispatch} to dynamically select the correct algorithm from the steps outlined above. Machine learning functions in 
 \textbf{TSML} are wrappers to the corresponding Scikit-Learn, Caret, and native Julia ML libraries. 
-There are more than hundred classifiers and regression functions available using a common API.
+There are more than hundreds of classifiers and regression functions available using \textbf{TSML}'s common API.
 
 \section{TSML Workflow}
 \label{sec:tsmlworkflow}
 %
 
-All major data types in \textbf{TSML} is a subtype of \emph{Transformer}. There are two major types of transformers, namely: \emph{filters} for data processing and \emph{learners} for machine learning. Both transformers implement \texttt{fit!} and \texttt{transform!} multidispatch functions. 
+All major data processing types in \textbf{TSML} are a subtype of the \texttt{Transformer}. There are two major types of transformers, namely: \emph{filters} for data processing and \emph{learners} for machine learning. Both transformers implement the \texttt{fit!} and \texttt{transform!} multidispatch functions. All filters are direct subtypes of the \texttt{Transformer} while all learners are subtypes of the \texttt{TSLearner}. The \texttt{TSLearner} is a direct subtype of the \texttt{Transformer}
+
 
 \vskip 6pt
 
 Filters are normally used for pre-processing tasks such as: imputation, normalization, feature extraction, feature transformation, scaling, etc.
-Consequently, filter's \texttt{fit!} and \texttt{transform!} functions expect one argument which represents an input data for feature processing. Each data type must implement \texttt{fit!} and \texttt{transform!} although in many cases of preprocessing data, only \texttt{transform!} operation is needed. The reason for this is explained below in the discussion of \emph{Pipeline} implementation. All filters are direct subtypes of \emph{Transformer}.
+Consequently, filter's \texttt{fit!} and \texttt{transform!} functions expect one argument which represents an input data for feature processing. Each data type must implement \texttt{fit!} and \texttt{transform!} although in some cases, only \texttt{transform!} operation is needed. For instance, square-root or log-transform filters do not require any initial computation of parameters to transform their inputs. On the other hand, feature transformations such as scaling, normalization, PCA, ICA, etc. require initial computation of certain parameters in its input before applying it to new datasets. The initial computations of these parameters are performed by the \texttt{fit!} function while their applications to new datasets are done by the \texttt{transform!} function. 
 
 \vskip 6pt
 
-On the other hand, a \emph{learner} requires learning of parameters between input and output. The training part to learn the parameters is handled by \texttt{fit!} function and \texttt{transform!} for the prediction. All learners are subtypes of \emph{TSLearner <: Transformer}. 
[…]
