### 2019-10-22 328dc58
grammar and layout edits
