### 2019-10-22 #51 @ppalmes
hi @matbesancon, here's the article proof  based on your suggestions.
