### 2019-10-23 #51 @ppalmes
Hi @matbesancon , let me know if the changes are ok.
