+++ paper/paper.tex
@@ -75,7 +75,7 @@ All major data processing types in \textbf{TSML} are subtypes of the \texttt{Tra
 \vskip 6pt
 
 Filters are normally used for pre-processing tasks such as imputation, normalization, feature extraction, feature transformation, scaling, etc.
-Consequently, filter's \texttt{fit!} and \texttt{transform!} functions expect one argument which represents an input data for feature extraction or transformation. Each data type must implement \texttt{fit!} and \texttt{transform!} although in some cases, only \texttt{transform!} operation is needed. For instance, \emph{square root} or \emph{log} filters do not require any initial computation of parameters to transform their inputs. On the other hand, feature transformations such as scaling, normalization, PCA, ICA, etc. require initial computation of certain parameters in their input before applying the transformation to new datasets. In these cases, initial computations of these parameters are performed by the \texttt{fit!} function while their applications to new datasets are done by the \texttt{transform!} function. 
+Consequently, filters' \texttt{fit!} and \texttt{transform!} functions expect one argument which represents an input data for feature extraction or transformation. Each data type must implement \texttt{fit!} and \texttt{transform!} although in some cases, only \texttt{transform!} operation is needed. For instance, \emph{square root} or \emph{log} filters do not require any initial computation of parameters to transform their inputs. On the other hand, feature transformations such as scaling, normalization, PCA, ICA, etc. require initial computation of certain parameters in their input before applying the transformation to new datasets. In these cases, initial computations of these parameters are performed by the \texttt{fit!} function while their applications to new datasets are done by the \texttt{transform!} function. 
 
 \vskip 6pt
 
