### 2019-10-24 #51 @ppalmes
Hi @matbesancon, I think pluralized is ok because I'm talking about all the filters.
