### 2019-10-24 #51 @matbesancon
ok, this paper looks good, thanks @christianpeel, @theogf!
