### 2019-10-29 #51 @ppalmes
@matbesancon, @christianpeel, @theogf

thanks all! highly appreciate the suggestions and corrections.
