### 2019-10-29 7136cf4
minor grammar edits
