+++ paper/paper.tex
@@ -84,7 +84,7 @@ Learners, on the other hand, expect two arguments (input vs output) and require 
 \vskip 6pt
 
 The \textbf{TSML} workflow borrows the idea of the  \emph{Unix} pipeline\cite{orchestra2014, combineml2016}. 
-The \texttt{Pipeline} data type is also a subtype of the \texttt{Transformer} and expects two arguments: input and output. The main elements in a \textbf{TSML} pipeline are series of transformers with each performing one specific task and does it well. The series of filters are used to perform pre-processing of the input while a machine learner at the end of the pipeline is used to learn the \emph{input--output} mapping. From the perspective of using the \texttt{Pipeline} where the last component is a machine learner, the \texttt{fit!} function is the training phase while the \texttt{transform!} function the prediction or classification phase.
+The \texttt{Pipeline} data type is also a subtype of the \texttt{Transformer} and expects two arguments: input and output. The main elements in a \textbf{TSML} pipeline are series of transformers with each performing one specific task and does it well. The series of filters are used to perform pre-processing of the input while a machine learner at the end of the pipeline is used to learn the \emph{input--output} mapping. From the perspective of using the \texttt{Pipeline} where the last component is a machine learner, the \texttt{fit!} function is the training phase while the \texttt{transform!} function is the prediction or classification phase.
 \vskip 6pt
 
 The \texttt{fit!} function in the \texttt{Pipeline} iteratively calls the \texttt{fit!} and \texttt{transform!} functions in a series of transformers. If the last transformer in the pipeline is a learner, the last transformed output from a series of filters will be used as input features for the \texttt{fit!} or training phase of the said learner.
