### 2019-10-29 #51 @matbesancon
i think there is an error in the YAML file, the affiliation is broken
