### 2019-10-29 #51 @ppalmes
@matbesancon, i haven't touched or edited paper.yml. my last git commit editing it was Aug 27. I don't know why this error appeared recently when my paper.yml used to work before. The affiliation stub is there.
