### 2019-10-29 81c75d3
fix paper.yml adding affiliation index
