### 2019-10-29 #51 @ppalmes
@matbesancon , apparently there was changes to the parsing of yml. I added affiliation index which now works but since all of us are in the same company, it doesn't make sense to have this index on in the affiliation.
