### 2019-10-30 #51 @matbesancon
perfect thanks. Yes that's just the way the YAML is parsed, otherwise it adds special cases
