### 2019-11-05 b898efa
binder notebooks support
