### 2019-11-06 #51 @ppalmes
@matbesancon , when is the expected date of publication? thanks.
