### 2019-11-11 eddf11f
add support for multicolumn imputation.
DateValMultiNNer is a filter that imputes multicolum dataframe. The
first column should be a Date class and the remaining columns should
be real numbers.
