+++ Project.toml
@@ -13,6 +13,7 @@ DecisionTree = "7806a523-6efd-50cb-b5f6-3fa6f1930dbb"
 FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
 GR = "28b8d3ca-fb5f-59d9-8090-bfdbd6d07a71"
 HTTP = "cd3eb016-35fb-5094-929b-558a96fad6f3"
+Impute = "f7bf1975-0170-51b9-8c5f-a992d46b9575"
 Interact = "c601a237-2ae4-5e1e-952c-7a85b0c7eef1"
 IterTools = "c8e1da08-722c-5040-9ed9-7db0dc04731e"
 JSON2 = "2535ab7d-5cd8-5a07-80ac-9b1792aadce3"
@@ -27,23 +28,23 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 
 [compat]
-julia = "1"
-CSV = "0.4, 0.5"
-Interact = "0.9, 0.10"
-GR = "0.35, 0.36, 0.37, 0.38, 0.39, 0.40, 0.41"
-JSON2 = "0.2, 0.3"
-HTTP = "0.6, 0.7, 0.8"
-IterTools = "1.1, 1.2"
-MLDataUtils = "0.2, 0.3, 0.4, 0.5"
-StatsBase = "0.25, 0.26, 0.27, 0.28, 0.29, 0.30, 0.31, 0.32"
-DecisionTree = "0.8, 0.9"
-Plots = "0.25, 0.26, 0.27"
-CodecBzip2 = "0.4, 0.5, 0.6"
-FileIO = "1.0"
-DataFrames = "0.17, 0.18, 0.19"
-Reexport = "0.1, 0.2"
-MLBase = "0.6, 0.7, 0.8"
 ArgParse = "0.5, 0.6"
+CSV = "0.4, 0.5"
+CodecBzip2 = "0.4, 0.5, 0.6"
+DataFrames = "0.17, 0.18, 0.19"
+DecisionTree = "0.8, 0.9"
+FileIO = "1.0"
+GR = "0.35, 0.36, 0.37, 0.38, 0.39, 0.40, 0.41"
+HTTP = "0.6, 0.7, 0.8"
+Interact = "0.9, 0.10"
+IterTools = "1.1, 1.2"
+JSON2 = "0.2, 0.3"
+MLBase = "0.6, 0.7, 0.8"
+MLDataUtils = "0.2, 0.3, 0.4, 0.5"
+Plots = "0.25, 0.26, 0.27"
+Reexport = "0.1, 0.2"
+StatsBase = "0.25, 0.26, 0.27, 0.28, 0.29, 0.30, 0.31, 0.32"
+julia = "1"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
