### 2019-11-11 e0bb5d0
add linear interpolation imputer
replaces missings with linear interpolation using Impute package.
