+++ Project.toml
@@ -36,6 +36,7 @@ DecisionTree = "0.8, 0.9"
 FileIO = "1.0"
 GR = "0.35, 0.36, 0.37, 0.38, 0.39, 0.40, 0.41"
 HTTP = "0.6, 0.7, 0.8"
+Impute = "0.1,0.2,0.3"
 Interact = "0.9, 0.10"
 IterTools = "1.1, 1.2"
 JSON2 = "0.2, 0.3"
