+++ README.md
@@ -1,183 +1,0 @@
-<div align="center"> <img
-src="https://ibm.github.io/TSML.jl/tsmllogo/tsmllogo13.png"
-alt="TSML Logo" width="250"></img> </div>
-
-| **Documentation** | **Build Status** | **Help** |
-|:---:|:---:|:---:|
-| [![][docs-dev-img]][docs-dev-url] [![][docs-stable-img]][docs-stable-url] | [![][travis-img]][travis-url] [![][codecov-img]][codecov-url] | [![][slack-img]][slack-url] [![][gitter-img]][gitter-url] |
-
-### TSML is a package for time series data processing, classification, clustering, and prediction written in [Julia](http://julialang.org/).
-
-The design/framework of this package is influenced heavily by Samuel Jenkins' [Orchestra.jl](https://github.com/svs14/Orchestra.jl) and Paulito Palmes [CombineML.jl](https://github.com/ppalmes/CombineML.jl) packages.
-
-Follow these links for a quick [Jupyter Notebook TSML Demo](https://github.com/IBM/TSML.jl/blob/master/docs/StaticPlotting.jl.ipynb)
-and the [JuliaCon 2019 Proceedings Paper](https://github.com/IBM/TSML.jl/blob/master/docs/JuliaCon2019Paper.pdf).
-
-## Package Features
-
-- TS data type clustering/classification for automatic data discovery
-- TS aggregation based on date/time interval
-- TS imputation based on `symmetric` Nearest Neighbors
-- TS statistical metrics for data quality assessment
-- TS ML wrapper with more than 100+ libraries from caret, scikitlearn, and julia
-- TS date/value matrix conversion of 1-D TS using sliding windows for ML input
-- Common API wrappers for ML libs from JuliaML, PyCall, and RCall
-- Pipeline API allows high-level description of the processing workflow
-- Specific cleaning/normalization workflow based on data type
-- Automatic selection of optimised ML model
-- Automatic segmentation of time-series data into matrix form for ML training and  prediction
-- Easily extensible architecture by using just two main interfaces: fit and transform
-- Meta-ensembles for robust prediction
-- Support for threads and distributed computation for scalability, and speed
-
-## Installation
-TSML is in the Julia Official package registry. The latest release can be installed at the Julia prompt using Julia's package management which is triggered by pressing `]` at the julia prompt:
-
-```julia
-julia> ]
-(v1.1) pkg> add TSML
-```
-
-Or, equivalently, via the `Pkg` API:
-
-```julia
-julia> using Pkg
-julia> Pkg.add("TSML")
-```
-
-## Documentation
-
-- [**STABLE**][docs-stable-url] &mdash; **documentation of the most recently tagged version.**
-- [**DEVEL**][docs-dev-url] &mdash; *documentation of the in-development version.*
-
-## Project Status
-
-TSML is tested and actively developed on Julia `1.0` and above for Linux and macOS.
-
-There is no support for Julia versions `0.4`, `0.5`, `0.6` and `0.7`.
-
-## Overview
-
-TSML (Time Series Machine Learning) is a package for Time Series data processing, classification, and prediction. It combines ML libraries from Python's ScikitLearn, R's Caret, and Julia using a common API and allows seamless ensembling and integration of heterogenous ML libraries to create complex models for robust time-series prediction.
-
-## Motivations
-Over the past years, the industrial sector has seen
-many innovations brought about by automation.
-Inherent in this automation is the installation of
-sensor networks for status monitoring and data collection.
-One of the major challenges in these data-rich
-environments is how to extract and exploit
-information from these large volume of data to
-detect anomalies, discover patterns to reduce
[…]
