+++ data/realdatatsclassification/model/README.md
@@ -1,1 +1,0 @@
-Stores model and stat dataframe extracted from time series datasets use for training the model.
