+++ Project.toml
@@ -11,3 +11,4 @@ Latexify = "23fbe1c1-3f47-55db-b15f-69d7ec21a316"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 TSML = "198dc43e-9e51-5cd7-9d40-d9794d335912"
+WebIO = "0f1e0344-ec1d-5b48-a673-e5cf874b6c29"
