### 2019-11-12 6e1f998
simplify module loading
