+++ Project.toml
@@ -34,7 +34,7 @@ CodecBzip2 = "0.4, 0.5, 0.6"
 DataFrames = "0.17, 0.18, 0.19"
 DecisionTree = "0.8, 0.9"
 FileIO = "1.0"
-GR = "0.35, 0.36, 0.37, 0.38, 0.39, 0.40, 0.41"
+GR = "0.35, 0.36, 0.37, 0.38, 0.39, 0.40, 0.41, 0.42"
 HTTP = "0.6, 0.7, 0.8"
 Impute = "0.1,0.2,0.3"
 Interact = "0.9, 0.10"
