### 2019-11-12 67f6877
less strict version number deps
