+++ Project.toml
@@ -1,7 +1,7 @@
 name = "TSML"
 uuid = "198dc43e-9e51-5cd7-9d40-d9794d335912"
 authors = ["Paulito Palmes <ppalmes@gmail.com>"]
-version = "2.4.0"
+version = "2.4.1"
 
 [deps]
 ArgParse = "c7e460c6-2fb9-53a9-8c5b-16f535851c63"
@@ -28,24 +28,24 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 
 [compat]
-ArgParse = "0.5, 0.6"
-CSV = "0.4, 0.5"
-CodecBzip2 = "0.4, 0.5, 0.6"
-DataFrames = "0.17, 0.18, 0.19"
-DecisionTree = "0.8, 0.9"
-FileIO = "1.0"
-GR = "0.35, 0.36, 0.37, 0.38, 0.39, 0.40, 0.41, 0.42"
-HTTP = "0.6, 0.7, 0.8"
-Impute = "0.1,0.2,0.3"
-Interact = "0.9, 0.10"
-IterTools = "1.1, 1.2"
-JSON2 = "0.2, 0.3"
-MLBase = "0.6, 0.7, 0.8"
-MLDataUtils = "0.2, 0.3, 0.4, 0.5"
-Plots = "0.25, 0.26, 0.27"
-Reexport = "0.1, 0.2"
-StatsBase = "0.25, 0.26, 0.27, 0.28, 0.29, 0.30, 0.31, 0.32"
-julia = "1"
+ArgParse = "≥ 0.5"
+CSV = "≥ 0.4"
+CodecBzip2 = "≥ 0.4"
+DataFrames = "≥ 0.17"
+DecisionTree = "≥ 0.8"
+FileIO = "≥ 1.0"
+GR = "≥ 0.35"
+HTTP = "≥ 0.6"
+Impute = "≥ 0.1"
+Interact = "≥ 0.9"
+IterTools = "≥ 1.1"
+JSON2 = "≥ 0.2"
+MLBase = "≥ 0.6"
+MLDataUtils = "≥ 0.2"
+Plots = "≥ 0.25"
+Reexport = "≥ 0.1"
+StatsBase = "≥ 0.25"
+julia = "≥ 1.0.0"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
