### 2019-11-14 f8b1b36
clone Impute.jl and remove init
