### 2019-11-26 ec7b21b
add upper bound in compat
