+++ Project.toml
@@ -29,22 +29,22 @@ Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 
 [compat]
 julia = "1"
-CSV = "≤ 1.0"
-Interact = "≤ 1.0"
-GR = "≤ 1.0"
-JSON2 = "≤ 1.0"
-HTTP = "≤ 1.0"
-IterTools = "^1.1"
-MLDataUtils = "≤ 1.0"
-StatsBase = "≤ 1.0"
-DecisionTree = "≤ 1.0"
-Plots = "≤ 1.0"
-CodecBzip2 = "≤ 1.0"
-FileIO = "^1.0"
-DataFrames = "≤ 1.0"
-Reexport = "≤ 1.0"
-MLBase = "≤ 1.0"
-ArgParse = "≤ 1.0"
+CSV = "0.4, 0.5"
+Interact = "0.9, 0.10"
+GR = "0.35, 0.36, 0.37, 0.38, 0.39, 0.40, 0.41, 0.42, 0.43, 0.44, 1.0"
+JSON2 = "0.2, 0.3"
+HTTP = "0.6, 0.7, 0.8, 0.9"
+IterTools = "1.1, 1.2, 1.3, 2.0"
+MLDataUtils = "0.2, 0.3, 0.4, 0.5, 1.0"
+StatsBase = "0.25, 0.26, 0.27, 0.28, 0.29, 0.30, 0.31, 0.32, 1.0"
+DecisionTree = "0.8, 0.9, 1.0"
+Plots = "0.25, 0.26, 0.27, 1.0"
+CodecBzip2 = "0.4, 0.5, 0.6, 1.0"
+FileIO = "1.0, 1.1"
+DataFrames = "0.17, 0.18, 0.19, 0.20, 1.0"
+Reexport = "0.1, 0.2, 0.3"
+MLBase = "0.6, 0.7, 0.8, 0.9, 1.0"
+ArgParse = "0.5, 0.6, 0.7, 1.0"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
