### 2019-11-27 #51 @ppalmes
hi @matbesancon, any news about publication date? thanks!
