### 2019-11-27 0a94b3b
CompatHelper: bump compat for "Plots" to "0.28"
