### 2019-11-27 113d8bc
CompatHelper: add new compat entry for "Tables" at version "0.2"
