+++ Project.toml
@@ -38,13 +38,14 @@ IterTools = "1.1, 1.2, 1.3, 2.0"
 MLDataUtils = "0.2, 0.3, 0.4, 0.5, 1.0"
 StatsBase = "0.25, 0.26, 0.27, 0.28, 0.29, 0.30, 0.31, 0.32, 1.0"
 DecisionTree = "0.8, 0.9, 1.0"
-Plots = "0.25, 0.26, 0.27, 1.0"
+Plots = "0.25, 0.26, 0.27, 0.28, 1.0"
 CodecBzip2 = "0.4, 0.5, 0.6, 1.0"
 FileIO = "1.0, 1.1"
 DataFrames = "0.17, 0.18, 0.19, 0.20, 1.0"
 Reexport = "0.1, 0.2, 0.3"
 MLBase = "0.6, 0.7, 0.8, 0.9, 1.0"
 ArgParse = "0.5, 0.6, 0.7, 1.0"
+Tables = "0.2"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
