### 2019-11-28 8e5fa30
CompatHelper: add new compat entry for "Revise" at version "2.3"
