+++ Project.toml
@@ -46,6 +46,7 @@ MLDataUtils = "0.2, 0.3, 0.4, 0.5, 1.0"
 MultivariateStats = "0.5, 0.6, 0.7, 0.8, 0.9, 1.0"
 Plots = "0.25, 0.26, 0.27, 0.28, 1.0"
 Reexport = "0.1, 0.2, 0.3"
+Revise = "2.3"
 StatsBase = "0.25, 0.26, 0.27, 0.28, 0.29, 0.30, 0.31, 0.32, 1.0"
 Tables = "0.2"
 julia = "1"
