### 2019-11-28 eb476f2
fix docs in normalizer; add schema support
