+++ test/test_normalizer.jl
@@ -42,7 +42,15 @@ function test_normalizer()
   norm = Normalizer(Dict(:method => :ppca))
   fit!(norm,X)
   res = transform!(norm,X)
-  @assert isapprox(std(res[:,1]),0.00408,atol=1e-2)
+  @test isapprox(std(res[:,1]),0.00408,atol=1e-2)
+  norm = Normalizer(Dict(:method => :log))
+  fit!(norm,X)
+  res = transform!(norm,X)
+  @test isapprox(std(res[:,1]),0.99941,atol=1e-2)
+  norm = Normalizer(Dict(:method => :sqrt))
+  fit!(norm,X)
+  res = transform!(norm,X)
+  @test isapprox(std(res[:,1]),0.2355,atol=1e-2)
 end
 @testset "Normalizer: zscore, unitrange, pca, ppca, fa" begin
   test_normalizer()
