+++ Project.toml
@@ -16,6 +16,7 @@ HTTP = "cd3eb016-35fb-5094-929b-558a96fad6f3"
 Interact = "c601a237-2ae4-5e1e-952c-7a85b0c7eef1"
 IterTools = "c8e1da08-722c-5040-9ed9-7db0dc04731e"
 JSON2 = "2535ab7d-5cd8-5a07-80ac-9b1792aadce3"
+JuliaDB = "a93385a2-3734-596a-9a66-3cfbb77141e6"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 MLBase = "f0e99cf1-93fa-52ec-9ecc-5026115318e0"
 MLDataUtils = "cc2ba9b6-d476-5e6d-8eaf-a92d5412d41d"
@@ -23,7 +24,6 @@ MultivariateStats = "6f286f6a-111f-5878-ab1e-185364afe411"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
-Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
 Serialization = "9e88b42a-f829-5b0c-bbe9-9e923198166b"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
