### 2019-11-28 d1a9ea4
add julia 1.3 travis test
