### 2019-11-29 1324ce1
add schema-based feature extraction
