+++ src/TSML.jl
@@ -2,11 +2,8 @@ module TSML
 using Reexport
 
 # reexport common functions to Main
-@reexport using CSV
-@reexport using Dates
-@reexport using DataFrames
-@reexport using Random
-@reexport using Statistics
+include("pkgdeps.jl")
+using .PkgDeps
 
 include("imputer/Imputers.jl")
 using .Imputers
