### 2019-11-29 e138ff3
fix pkg deps export namespaces
