+++ src/pkgdeps.jl
@@ -1,15 +1,15 @@
 @reexport module PkgDeps
 using CSV: read, write
 using Dates: DateTime, Date,Dates, Day,Year, Month, Week, Minute, Hour, year, day, month, hour, minute, week, dayofweek, dayofquarter
-using DataFrames: DataFrame, DataFrameRow, DataFrameRows, DataFrameColumns, nrow, ncol
+using DataFrames: DataFrame, DataFrameRow, DataFrameRows, DataFrameColumns, nrow, ncol, eachcol
 using Random: seed!
 using Statistics: mean, median, middle, quantile, std, var, cor, cov
-import Random
+import Random, CSV, Dates, DataFrames, Statistics
 
-export Random
+export Random, CSV, Dates, DataFrames, Statistics
 export read, write
 export DateTime, Date,Dates, Day,Year, Month, Week, Minute, Hour, year, day, month, hour, minute, week, dayofweek, dayofquarter
-export DataFrame, DataFrameRow, DataFrameRows, DataFrameColumns, nrow, ncol
+export DataFrame, DataFrameRow, DataFrameRows, DataFrameColumns, nrow, ncol, eachcol
 export seed!
 export mean, median, middle, quantile, std, var, cor, cov
 
