### 2019-11-29 ca6f141
CompatHelper: add new compat entry for "JuliaDB" at version "0.13"
