### 2019-11-29 11ff41b
remove 1.3 support due to lazy arrays CSV bug
