+++ .travis.yml
@@ -7,7 +7,6 @@ julia:
   - 1.0
   - 1.1
   - 1.2
-  - 1.3
   - nightly
 env:
   - JULIA_NUM_THREADS=8
