### 2019-11-29 68d8c64
showtree to see structure of pipeline and transformers
