### 2019-11-30 657e145
add juliadb version compat
