+++ Project.toml
@@ -42,6 +42,7 @@ HTTP = "0.6, 0.7, 0.8, 0.9"
 Interact = "0.9, 0.10"
 IterTools = "1.1, 1.2, 1.3, 2.0"
 JSON2 = "0.2, 0.3"
+JuliaDB = "0.13"
 MLBase = "0.6, 0.7, 0.8, 0.9, 1.0"
 MLDataUtils = "0.2, 0.3, 0.4, 0.5, 1.0"
 MultivariateStats = "0.5, 0.6, 0.7, 0.8, 0.9, 1.0"
