### 2019-11-30 ab88c93
CompatHelper: add new compat entry for "AbstractTrees" at version "0.2"
