+++ Project.toml
@@ -31,6 +31,7 @@ StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 
 [compat]
+AbstractTrees = "0.2"
 ArgParse = "0.5, 0.6, 0.7, 1.0"
 CSV = "0.4, 0.5"
 CodecBzip2 = "0.4, 0.5, 0.6, 1.0"
