+++ .travis.yml
@@ -5,8 +5,7 @@ os:
   - osx
 julia:
   - 1.0
-  - 1.1
-  - 1.2
+  - 1.3
   - nightly
 env:
   - JULIA_NUM_THREADS=8
