+++ Project.toml
@@ -1,7 +1,7 @@
 name = "TSML"
 uuid = "198dc43e-9e51-5cd7-9d40-d9794d335912"
 authors = ["Paulito Palmes <ppalmes@gmail.com>"]
-version = "2.4.2"
+version = "2.4.3"
 
 [deps]
 AbstractTrees = "1520ce14-60c1-5f80-bbc7-55ef81b5835c"
@@ -36,7 +36,7 @@ ArgParse = "0.5, 0.6, 0.7, 1.0"
 CSV = "0.4, 0.5"
 CodecBzip2 = "0.4, 0.5, 0.6, 1.0"
 DataFrames = "0.17, 0.18, 0.19, 0.20, 1.0"
-DecisionTree = "0.8, 0.9, 1.0, 0.10"
+DecisionTree = "0.8, 0.9, 0.10"
 FileIO = "1.0, 1.1"
 GR = "0.35, 0.36, 0.37, 0.38, 0.39, 0.40, 0.41, 0.42, 0.43, 0.44, 1.0"
 HTTP = "0.6, 0.7, 0.8, 0.9"
