### 2019-12-12 af0f619
fix deprecations, move tree visualization to extra
