### 2019-12-12 cda2e33
fix tsmfullstat arg type to abstract arrays
