### 2019-12-12 d179760
collect lazy arrays for skipmissing application
