### 2019-12-12 48bd51b
disable statifier test due to buggy LazyArray
