### 2019-12-26 7ba91b6
fix normalizers and preprocessing functions
