+++ test/runtests.jl
@@ -13,6 +13,7 @@ include("test_outliernicer.jl")
 include("test_plotter.jl")
 include("test_normalizer.jl")
 include("test_schemalizer.jl")
+include("test_crossvalidator.jl")
 
 #include("test_timescaledb.jl")
 
