### 2019-12-30 dc04ee9
simplify cross-validation
