+++ src/crossvalidator.jl
@@ -1,31 +1,27 @@
 @reexport module CrossValidators
 
-using TSML: TSMLTypes, Utils, Pipeline,fit!,transform!, kfold
+using TSML.TSMLTypes
+using TSML.Utils
 using Statistics: mean, std
 using DataFrames
 
 export crossvalidate
 
-function crossvalidate(pl::Pipeline,data::DataFrame,
-		       cfndx::Array,
-		       ctfndx::Array,
-		       classndx::Integer,
+function crossvalidate(pl::Transformer,X::Union{DataFrame,Matrix},Y::Vector,
 		       pfunc::Function,nfolds=10) 
   ## flatten arrays
+  @assert size(X)[1] == length(Y)
   ppl = deepcopy(pl)
-  contfndx = cfndx|> Iterators.flatten |> collect
-  catfndx = ctfndx|> Iterators.flatten |> collect
-  featuresndx = merge(contfndx,catfndx)
-  input = data[:,featuresndx]
-  output = data[:,classndx]
-  rowsize = size(data)[1]
+  input = X |> Matrix
+  output = Y
+  rowsize = size(input)[1]
   folds = kfold(rowsize,nfolds)
   pacc = Float64[]
   for (fold,trainndx) in enumerate(folds)
     testndx = setdiff(1:rowsize,trainndx)
-    trX = input[trainndx,:] |> Matrix
+    trX = input[trainndx,:] 
     trY = output[trainndx] |> collect
-    tstX = input[testndx,:] |> Matrix
+    tstX = input[testndx,:]
     tstY = output[testndx] |> collect
     res = pipe_accuracy(ppl,pfunc,trX,trY,tstX,tstY)
     push!(pacc,res)
