### 2019-12-31 48c670a
refactor and remove reexport macro
