+++ src/TSML.jl
@@ -1,5 +1,6 @@
 module TSML
-using Reexport
+
+export fit!, transform!
 
 # reexport common functions to Main
 include("pkgdeps.jl")
@@ -10,63 +11,90 @@ using .Imputers
 
 include("types.jl")
 using .TSMLTypes
+export  Transformer,TSLearner
 
 include("utils.jl")
 using .Utils
+export holdout, kfold, score, infer_eltype, nested_dict_to_tuples, 
+       nested_dict_set!, nested_dict_merge, create_transformer,
+       mergedict, getiris,
+       skipmean,skipmedian,skipstd,
+       aggregatorclskipmissing
 
 include("baseline.jl")
 using .BaselineAlgos
+export Baseline,Identity
 
 include("basefilters.jl")
 using .BaseFilters
+export Imputer,Pipeline,OneHotEncoder,Wrapper
 
 include("valdatefilters.jl")
 using .ValDateFilters
+export Matrifier,Dateifier,
+       DateValizer,DateValgator,DateValNNer,DateValMultiNNer,
+       CSVDateValReader, CSVDateValWriter, DateValLinearImputer,
+       BzCSVDateValReader
 
 include("statifier.jl")
 using .Statifiers
+export Statifier,tsmlfullstat
 
 include("mlbase.jl")
 using .MLBaseWrapper
+export Standardize,standardize, standardize!, 
+       transform, estimate, transform,StandardScaler
 
 include("decisiontree.jl")
 using .DecisionTreeLearners
+export PrunedTree,RandomForest,Adaboost
 
 include("normalizer.jl")
 using .Normalizers
-
+export Normalizer
 
 include("monotonicer.jl")
 using .Monotonicers
+export Monotonicer,ismonotonic,dailyflips
 
 include("cliwrapper.jl")
 using .CLIWrappers
+export tsmlrun
 
 include("tsclassifier.jl")
 using .TSClassifiers
+export TSClassifier, getstats
 
 include("outliernicer.jl")
 using .Outliernicers
+export Outliernicer
 
 include("plotter.jl")
 using .Plotters
+export Plotter
 
 include("timescaledb.jl")
 using .TimescaleDBs
+export TimescaleDB
 
 include("demo.jl")
 using .TSMLDemo
-
-include("argparse.jl")
-using .ArgumentParsers
+export tsml_demo
 
 include("ensemble.jl")
 using .EnsembleMethods
+export VoteEnsemble, StackEnsemble, BestLearner
 
 include("schema.jl")
 using .Schemalizers
+export Schemalizer, ML, table
 
 include("crossvalidator.jl")
 using .CrossValidators
+export crossvalidate
+
+include("argparse.jl")
+using .ArgumentParsers
+export tsmlmain
 
 end # module
