+++ src/basefilters.jl
@@ -1,4 +1,4 @@
-@reexport module BaseFilters
+module BaseFilters
 
 using Dates
 using DataFrames
