### 2019-12-31 23fbebf
lessen lower bound for crossvalidation performance
