+++ test/test_crossvalidator.jl
@@ -11,15 +11,15 @@ function test_crossvalidator()
   X=data[:,1:4] 
   Y=data[:,5] |> Vector{String}
   rf = RandomForest()
-  @test crossvalidate(rf,X,Y,acc).mean ≈ 94.0
+  @test crossvalidate(rf,X,Y,acc).mean > 80.0
   Random.seed!(123)
   ppl1 = Pipeline(Dict(:transformers=>[RandomForest()]))
-  @test crossvalidate(ppl1,X,Y,acc).mean ≈ 94.0
+  @test crossvalidate(ppl1,X,Y,acc).mean > 80.0
   Random.seed!(123)
   ohe = OneHotEncoder()
   stdsc= StandardScaler()
   ppl2 = Pipeline(Dict(:transformers=>[ohe,stdsc,RandomForest()]))
-  @test crossvalidate(ppl2,X,Y,acc).mean ≈ 94.66666667
+  @test crossvalidate(ppl2,X,Y,acc).mean > 80.0
   Random.seed!(123)
   mpca = Normalizer(Dict(:method=>:pca))
   mppca = Normalizer(Dict(:method=>:ppca))
@@ -33,7 +33,7 @@ function test_crossvalidator()
   @test size(transform!(ppl3,X))[1] == length(Y)
   Random.seed!(123)
   ppl5 = Pipeline(Dict(:transformers=>[msqrt,mlog,mppca,RandomForest()]))
-  @test crossvalidate(ppl5,X,Y,acc).mean ≈ 77.333333
+  @test crossvalidate(ppl5,X,Y,acc).mean > 50.0
 end
 @testset "CrossValidator" begin
   test_crossvalidator()
