+++ src/imputer/imputors/fill.jl
@@ -12,19 +12,6 @@ be handled independently.
   summary information
 
 # Example
-```jldoctest
-julia> using Impute: Fill, Context, impute
-
-julia> M = [1.0 2.0 missing missing 5.0; 1.1 2.2 3.3 missing 5.5]
-2×5 Array{Union{Missing, Float64},2}:
- 1.0  2.0   missing  missing  5.0
- 1.1  2.2  3.3       missing  5.5
-
-julia> impute(M, Fill(; context=Context(; limit=1.0)); dims=2)
-2×5 Array{Union{Missing, Float64},2}:
- 1.0  2.0  2.66667  2.66667  5.0
- 1.1  2.2  3.3      3.025    5.5
-```
 """
 struct Fill{T} <: Imputor
     value::T
