+++ src/imputer/imputors/chain.jl
@@ -22,15 +22,6 @@ Compose new `Imputor` chains with the composition operator
 
 # Example
 
-```jldoctest
-julia> using TSML.Imputers: impute, Interpolate, NOCB, LOCF, Context
-
-julia> ctx = Context(; limit=1.0)
-Context(0, 0, 1.0, ismissing, Impute.complete)
-
-julia> imp = Interpolate(; context=ctx) ∘ NOCB(; context=ctx) ∘ LOCF(; context=ctx)
-Impute.Chain(Impute.Imputor[Interpolate(2, Context(0, 0, 1.0, ismissing, complete)), NOCB(2, Context(0, 0, 1.0, ismissing, complete)), LOCF(2, Context(0, 0, 1.0, ismissing, complete))])
-```
 """
 Base.:(∘)(a::Imputor, b::Imputor) = Chain([a, b])
 function Base.:(∘)(a::Chain, b::Imputor)
