+++ src/imputer/imputors/interp.jl
@@ -14,19 +14,6 @@ that all missing values will be imputed.
   summary information
 
 # Example
-```jldoctest
-julia> using Impute: Interpolate, Context, impute
-
-julia> M = [1.0 2.0 missing missing 5.0; 1.1 2.2 3.3 missing 5.5]
-2×5 Array{Union{Missing, Float64},2}:
- 1.0  2.0   missing  missing  5.0
- 1.1  2.2  3.3       missing  5.5
-
-julia> impute(M, Interpolate(; context=Context(; limit=1.0)); dims=2)
-2×5 Array{Union{Missing, Float64},2}:
- 1.0  2.0  3.0  4.0  5.0
- 1.1  2.2  3.3  4.4  5.5
-```
 """
 struct Interpolate <: Imputor
     context::AbstractContext
