+++ src/imputer/imputors/locf.jl
@@ -17,19 +17,6 @@ that all missing values will be imputed.
   summary information
 
 # Example
-```jldoctest
-julia> using Impute: LOCF, Context, impute
-
-julia> M = [1.0 2.0 missing missing 5.0; 1.1 2.2 3.3 missing 5.5]
-2×5 Array{Union{Missing, Float64},2}:
- 1.0  2.0   missing  missing  5.0
- 1.1  2.2  3.3       missing  5.5
-
-julia> impute(M, LOCF(; context=Context(; limit=1.0)); dims=2)
-2×5 Array{Union{Missing, Float64},2}:
- 1.0  2.0  2.0  2.0  5.0
- 1.1  2.2  3.3  3.3  5.5
-```
 """
 struct LOCF <: Imputor
     context::AbstractContext
