+++ src/imputer/imputors.jl
@@ -93,19 +93,6 @@ if this is not the desired behaviour custom imputor methods should overload this
 * `AbstractMatrix`: the input `data` with values imputed
 
 # Example
-```jldoctest
-julia> using Impute: Interpolate, Context, impute
-
-julia> M = [1.0 2.0 missing missing 5.0; 1.1 2.2 3.3 missing 5.5]
-2×5 Array{Union{Missing, Float64},2}:
- 1.0  2.0   missing  missing  5.0
- 1.1  2.2  3.3       missing  5.5
-
-julia> impute(M, Interpolate(; context=Context(; limit=1.0)); dims=2)
-2×5 Array{Union{Missing, Float64},2}:
- 1.0  2.0  3.0  4.0  5.0
- 1.1  2.2  3.3  4.4  5.5
-```
 """
 function impute!(data::AbstractMatrix, imp::Imputor; dims=1)
     for var in varwise(data; dims=dims)
@@ -128,29 +115,6 @@ if this is not the desired behaviour custom imputor methods should overload this
 * the input `data` with values imputed
 
 # Example
-``jldoctest
-julia> using DataFrames; using Impute: Interpolate, Context, impute
-julia> df = DataFrame(:a => [1.0, 2.0, missing, missing, 5.0], :b => [1.1, 2.2, 3.3, missing, 5.5])
-5×2 DataFrame
-│ Row │ a        │ b        │
-│     │ Float64⍰ │ Float64⍰ │
-├─────┼──────────┼──────────┤
-│ 1   │ 1.0      │ 1.1      │
-│ 2   │ 2.0      │ 2.2      │
-│ 3   │ missing  │ 3.3      │
-│ 4   │ missing  │ missing  │
-│ 5   │ 5.0      │ 5.5      │
-
-julia> impute(df, Interpolate(; context=Context(; limit=1.0)))
-5×2 DataFrame
-│ Row │ a        │ b        │
-│     │ Float64⍰ │ Float64⍰ │
-├─────┼──────────┼──────────┤
-│ 1   │ 1.0      │ 1.1      │
-│ 2   │ 2.0      │ 2.2      │
-│ 3   │ 3.0      │ 3.3      │
-│ 4   │ 4.0      │ 4.4      │
-│ 5   │ 5.0      │ 5.5      │
 """
 function impute!(table, imp::Imputor)
     istable(table) || throw(MethodError(impute!, (table, imp)))
