### 2020-01-12 e241b25
CompatHelper: bump compat for "AbstractTrees" to "0.3"
