### 2020-01-24 b6cfd38
CompatHelper: bump compat for "GR" to "0.45"
