### 2020-01-27 ed4a697
bump compat versions and add crossvalidator module
