+++ Project.toml
@@ -1,7 +1,7 @@
 name = "TSML"
 uuid = "198dc43e-9e51-5cd7-9d40-d9794d335912"
 authors = ["Paulito Palmes <ppalmes@gmail.com>"]
-version = "2.4.4"
+version = "2.4.5"
 
 [deps]
 AbstractTrees = "1520ce14-60c1-5f80-bbc7-55ef81b5835c"
@@ -29,14 +29,14 @@ StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 
 [compat]
-AbstractTrees = "0.2"
+AbstractTrees = "0.2, 0.3"
 ArgParse = "0.5, 0.6, 0.7, 1.0"
 CSV = "0.4, 0.5"
 CodecBzip2 = "0.4, 0.5, 0.6, 1.0"
 DataFrames = "0.17, 0.18, 0.19, 0.20, 1.0"
 DecisionTree = "0.8, 0.9, 0.10, 1.0"
 FileIO = "1.0, 1.1"
-GR = "0.35, 0.36, 0.37, 0.38, 0.39, 0.40, 0.41, 0.42, 0.43, 0.44, 1.0"
+GR = "0.35, 0.36, 0.37, 0.38, 0.39, 0.40, 0.41, 0.42, 0.43, 0.44, 0.45, 1.0"
 HTTP = "0.6, 0.7, 0.8, 0.9"
 IterTools = "1.1, 1.2, 1.3, 2.0"
 JSON2 = "0.2, 0.3"
