### 2020-01-27 3f32baf
CompatHelper: bump compat for "GR" to "0.46"
