+++ Project.toml
@@ -36,7 +36,7 @@ CodecBzip2 = "0.4, 0.5, 0.6, 1.0"
 DataFrames = "0.17, 0.18, 0.19, 0.20, 1.0"
 DecisionTree = "0.8, 0.9, 0.10, 1.0"
 FileIO = "1.0, 1.1"
-GR = "0.35, 0.36, 0.37, 0.38, 0.39, 0.40, 0.41, 0.42, 0.43, 0.44, 0.45, 1.0"
+GR = "0.35, 0.36, 0.37, 0.38, 0.39, 0.40, 0.41, 0.42, 0.43, 0.44, 0.45, 1.0, 0.46"
 HTTP = "0.6, 0.7, 0.8, 0.9"
 IterTools = "1.1, 1.2, 1.3, 2.0"
 JSON2 = "0.2, 0.3"
