+++ Project.toml
@@ -30,22 +30,22 @@ Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 
 [compat]
 AbstractTrees = "0.2, 0.3"
-ArgParse = "0.5, 0.6, 0.7, 1.0"
+ArgParse = "0.5, 0.6, 0.7"
 CSV = "0.4, 0.5"
-CodecBzip2 = "0.4, 0.5, 0.6, 1.0"
-DataFrames = "0.17, 0.18, 0.19, 0.20, 1.0"
-DecisionTree = "0.8, 0.9, 0.10, 1.0"
+CodecBzip2 = "0.4, 0.5, 0.6"
+DataFrames = "0.17, 0.18, 0.19, 0.20"
+DecisionTree = "0.8, 0.9, 0.10"
 FileIO = "1.0, 1.1"
-GR = "0.35, 0.36, 0.37, 0.38, 0.39, 0.40, 0.41, 0.42, 0.43, 0.44, 0.45, 0.46, 1.0"
+GR = "0.35, 0.36, 0.37, 0.38, 0.39, 0.40, 0.41, 0.42, 0.43, 0.44, 0.45, 0.46"
 HTTP = "0.6, 0.7, 0.8, 0.9"
-IterTools = "1.1, 1.2, 1.3, 2.0"
+IterTools = "1.1, 1.2, 1.3"
 JSON2 = "0.2, 0.3"
 JuliaDB = "0.13"
-MLBase = "0.6, 0.7, 0.8, 0.9, 1.0"
-MLDataUtils = "0.2, 0.3, 0.4, 0.5, 1.0"
-MultivariateStats = "0.5, 0.6, 0.7, 0.8, 0.9, 1.0"
-Plots = "0.25, 0.26, 0.27, 0.28, 1.0"
-StatsBase = "0.25, 0.26, 0.27, 0.28, 0.29, 0.30, 0.31, 0.32, 1.0"
+MLBase = "0.6, 0.7, 0.8, 0.9"
+MLDataUtils = "0.2, 0.3, 0.4, 0.5"
+MultivariateStats = "0.5, 0.6, 0.7, 0.8, 0.9"
+Plots = "0.25, 0.26, 0.27, 0.28"
+StatsBase = "0.25, 0.26, 0.27, 0.28, 0.29, 0.30, 0.31, 0.32"
 Tables = "0.2"
 julia = "1"
 
