+++ Project.toml
@@ -44,7 +44,7 @@ JuliaDB = "0.13"
 MLBase = "0.6, 0.7, 0.8, 0.9"
 MLDataUtils = "0.2, 0.3, 0.4, 0.5"
 MultivariateStats = "0.5, 0.6, 0.7, 0.8, 0.9"
-Plots = "0.25, 0.26, 0.27, 0.28"
+Plots = "0.25, 0.26, 0.27, 0.28, 0.29"
 StatsBase = "0.25, 0.26, 0.27, 0.28, 0.29, 0.30, 0.31, 0.32"
 Tables = "0.2"
 julia = "1"
