### 2020-02-12 c405719
CompatHelper: bump compat for "Tables" to "1.0"
