+++ Project.toml
@@ -46,7 +46,7 @@ MLDataUtils = "0.2, 0.3, 0.4, 0.5"
 MultivariateStats = "0.5, 0.6, 0.7, 0.8, 0.9"
 Plots = "0.25, 0.26, 0.27, 0.28"
 StatsBase = "0.25, 0.26, 0.27, 0.28, 0.29, 0.30, 0.31, 0.32"
-Tables = "0.2"
+Tables = "0.2, 1.0"
 julia = "1"
 
 [extras]
