+++ Project.toml
@@ -30,7 +30,7 @@ Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 
 [compat]
 AbstractTrees = "0.2, 0.3"
-ArgParse = "0.5, 0.6, 0.7"
+ArgParse = "0.5, 0.6, 0.7, 1.1"
 CSV = "0.4, 0.5"
 CodecBzip2 = "0.4, 0.5, 0.6"
 DataFrames = "0.17, 0.18, 0.19, 0.20"
