### 2020-02-23 b8ffbaa
fix deprecations in StatsBase
