### 2020-02-26 6080a58
added linear and combo pipelines; feature selectors
