+++ src/featureselector.jl
@@ -1,0 +1,184 @@
+module FeatureSelectors
+
+using DataFrames
+using Random
+
+using TSML.TSMLTypes
+using TSML.BaseFilters
+using TSML.Utils
+
+import TSML.TSMLTypes: fit!, transform!
+export fit!, transform!
+export FeatureSelector, CatFeatureSelector, NumFeatureSelector, CatNumDiscriminator
+
+export feature_test
+
+# generic way to extract num/cat features by specifying their columns
+mutable struct FeatureSelector <: Transformer
+    name::String
+    model::Dict
+    args::Dict
+
+    function FeatureSelector(args::Dict = Dict())
+	default_args = Dict(
+			    :name => "featureselector",
+			    :columns => Int[],
+			    )
+	cargs=nested_dict_merge(default_args,args)
+	cargs[:name] = cargs[:name]*"_"*randstring(3)
+	new(cargs[:name],Dict(),cargs)
+    end
+end
+
+function FeatureSelector(cols::Vector{Int})
+    FeatureSelector(Dict(:columns => cols))
+end
+
+function fit!(ft::FeatureSelector, features::DataFrame, labels::Vector=[])
+    if features == DataFrame()
+	error("empty dataframe")
+    end
+    ft.model = ft.args
+end
+
+function transform!(ft::FeatureSelector, features::DataFrame)
+    if features == DataFrame()
+	error("empty dataframe")
+    end
+    return features[:,ft.model[:columns]]
+end
+
+# ----------
+# automatically extracts cat features based on their inferred element non-number types
+mutable struct CatFeatureSelector <: Transformer
+    name::String
+    model::Dict
+    args::Dict
+
+    function CatFeatureSelector(args::Dict = Dict())
+	default_args = Dict(
+			    :name => "catfeatsel",
+			    )
+	cargs=nested_dict_merge(default_args,args)
+	cargs[:name] = cargs[:name]*"_"*randstring(3)
+	new(cargs[:name],Dict(),cargs)
+    end
+end
+
+function fit!(ft::CatFeatureSelector, features::DataFrame, labels::Vector=[])
+    if features == DataFrame()
+        error("empty dataframe")
+    end
+    catcols,_ = find_catnum_columns(features)
+
+    # create model
+    ft.model = Dict(
+		    :nominal_columns => catcols
+		    )
+end
+
+function transform!(ft::CatFeatureSelector, features::DataFrame)
+    catcols = ft.model[:nominal_columns]
+    return features[:,catcols]
+end
+
+# ---------
+# automatically extracts numeric features based on their inferred element types
+mutable struct NumFeatureSelector <: Transformer
+    name::String
+    model::Dict
+    args::Dict
+
+    function NumFeatureSelector(args::Dict = Dict())
+	default_args = Dict(
+			    :name => "numfeatsel"
+			    )
+	cargs=nested_dict_merge(default_args,args)
+	cargs[:name] = cargs[:name]*"_"*randstring(3)
+	new(cargs[:name],Dict(),cargs)
+    end
+end
+
+function fit!(ft::NumFeatureSelector, features::DataFrame, labels::Vector=[])
+    if features == DataFrame()
+        error("empty dataframe")
+    end
+    _,realcols = find_catnum_columns(features)
+
+    # create model
+    ft.model = Dict(
+		    :real_columns => realcols
+		    )
+end
+
+function transform!(ft::NumFeatureSelector, features::DataFrame)
+    realcols = ft.model[:real_columns]
+    return features[:,realcols]
+end
+
+
+# ---------
[…]
