### 2020-02-26 68e34fc
fix infer of elements in utils
