### 2020-02-27 0dc16d0
CompatHelper: bump compat for "GR" to "0.47"
