### 2020-03-03 ce1e79a
deploydocs() fix and add ssh keys
