+++ README.md
@@ -176,7 +176,7 @@ Usage questions can be posted in:
 [docs-stable-img]: https://img.shields.io/badge/docs-stable-blue.svg
 [docs-stable-url]: https://ibm.github.io/TSML.jl/stable/
 [docs-dev-img]: https://img.shields.io/badge/docs-dev-blue.svg
-[docs-dev-url]: https://ibm.github.io/TSML.jl/latest/
+[docs-dev-url]: https://ibm.github.io/TSML.jl/dev/
 
 [travis-img]: https://travis-ci.org/IBM/TSML.jl.svg?branch=master
 [travis-url]: https://travis-ci.org/IBM/TSML.jl
