### 2020-03-05 b2dcafb
use symbolic pipeline in README, use |> in pipeline
