### 2020-03-05 3c2a629
improve README with example of modeling
