### 2020-03-05 fb91f52
edits in README
