### 2020-03-05 84d8a55
support for symbolic pipeline
