### 2020-03-05 3c5891b
use @pipeline in the docs
