+++ docs/src/tutorial/monotonic_plotting_pdf.md
@@ -17,12 +17,11 @@ We can use the `Plotter` filter to visualize the generated data.
 using TSML
 
 Random.seed!(123)
-pltr = Plotter(Dict(:interactive => false, :pdfoutput => true))
+pltr = Plotter(Dict(:interactive => false,:pdfoutput => true))
 mdates = DateTime(2017,12,1,1):Dates.Hour(1):DateTime(2017,12,31,10) |> collect
 mvals = rand(length(mdates)) |> cumsum
 df =  DataFrame(Date=mdates ,Value = mvals)
-fit!(pltr,df)
-transform!(pltr,df);
+fit_transform!(pltr,df)
 ```
 
 Now that we have a monotonic data, let's use the `Monotonicer` to normalize and plot the result:
@@ -31,13 +30,10 @@ Now that we have a monotonic data, let's use the `Monotonicer` to normalize and 
 using TSML
 
 mono = Monotonicer(Dict())
-pipeline = Pipeline(Dict(
-   :transformers => [mono,pltr]
-   )
-)
 
-fit!(pipeline,df)
-res=transform!(pipeline,df);
+pipeline = @pipeline mono |> pltr
+
+res=fit_transform!(pipeline,df)
 ```
 
 ## Real Data Example
@@ -68,6 +64,7 @@ valgator = DateValgator(Dict(:dateinterval=>Dates.Hour(1)))
 valnner = DateValNNer(Dict(:dateinterval=>Dates.Hour(1)))
 stfier = Statifier(Dict(:processmissing=>true))
 mono = Monotonicer(Dict())
+pltr = Plotter(Dict(:interactive => false))
 nothing #hide
 ```
 
@@ -78,23 +75,16 @@ Let's test by feeding the regular time series type to the pipeline. We expect th
 
 - Pipeline with `Monotonicer`: regular time series
 ```@example mono
-pipeline = Pipeline(Dict(
-    :transformers => [regularfilecsv,valgator,valnner,mono,pltr]
-   )
-)
-fit!(pipeline)
-transform!(pipeline);
+pipeline = @pipeline regularfilecsv |> valgator |> valnner |> mono |> pltr
+
+fit_transform!(pipeline)
 ```
 
 - Pipeline without `Monotonicer`: regular time series
 ```@example mono
-pipeline = Pipeline(Dict(
-    :transformers => [regularfilecsv,valgator,valnner,pltr]
-   )
-)
-fit!(pipeline)
-transform!(pipeline)
-nothing #hide
+pipeline = @pipeline regularfilecsv |> valgator |> valnner |> pltr
+
+fit_transform!(pipeline)
 ```
 
 Notice that the plots are the same with or without the `Monotonicer` instance.
@@ -104,23 +94,16 @@ Let's now feed the same pipeline with a monotonic csv data.
 
 - Pipeline without `Monotonicer`: monotonic time series
 ```@example mono
-pipeline = Pipeline(Dict(
-    :transformers => [monofilecsv,valgator,valnner,pltr]
-   )
-)
-fit!(pipeline)
-transform!(pipeline)
-nothing #hide
+pipeline = @pipeline monofilecsv |> valgator |> valnner |> pltr
+
+fit_transform!(pipeline)
 ```
 
 - Pipeline with `Monotonicer`: monotonic time series
 ```@example mono
-pipeline = Pipeline(Dict(
-    :transformers => [monofilecsv,valgator,valnner,mono,pltr]
-   )
-)
-fit!(pipeline)
-transform!(pipeline);
+pipeline = @pipeline monofilecsv |> valgator |> valnner |> mono |> pltr
+
+fit_transform!(pipeline)
 ```
 
 Notice that without the `Monotonicer` instance, the data is monotonic. Applying
[…]
