+++ .travis.yml
@@ -24,6 +24,8 @@ jobs:
     - julia: nightly
   include:
     - stage: "Documentation"
+      julia: 1.3
+      os: linux
       script:
         - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
         - julia --project=docs/ docs/make.jl
