### 2020-03-11 2108f54
remove redundant modules based on automlpipeline
