+++ src/cliwrapper.jl
@@ -1,10 +1,10 @@
 module CLIWrappers
 
-using TSML
-using TSML.TSMLTypes
-using TSML.ValDateFilters
-using TSML.Utils
+using AutoMLPipeline
+using AutoMLPipeline.AbsTypes
+using AutoMLPipeline.Utils
 
+using TSML.ValDateFilters
 using TSML: CSVDateValReader
 using TSML: CSVDateValWriter
 using TSML.Statifiers
@@ -35,12 +35,8 @@ function rawstat(inputname::AbstractString,outputname::AbstractString="",datefmt
     isfile(inputname) || error("input file name does not exist")
     csvreader = CSVDateValReader(Dict(:filename=>inputname,:dateformat=>datefmt))
     stfier = Statifier()
-    lpipe = Pipeline(Dict(
-        :transformers => [csvreader,stfier]
-      )
-    )
-    fit!(lpipe)
-    res=transform!(lpipe)
+    lpipe = @pipeline csvreader |> stfier
+    res=fit_transform!(lpipe)
     if outputname != ""
         res |> CSV.write(outputname)
     end
@@ -52,12 +48,8 @@ function aggregatedstat(inputname::AbstractString,outputname::AbstractString="",
     csvreader = CSVDateValReader(Dict(:filename=>inputname,:dateformat=>datefmt))
     valgator = DateValgator(COMMONARG)
     stfier = Statifier()
-    lpipe = Pipeline(Dict(
-        :transformers => [csvreader,valgator,stfier]
-      )
-    )
-    fit!(lpipe)
-    res=transform!(lpipe)
+    lpipe = @pipeline csvreader |> valgator |> stfier
+    res=fit_transform!(lpipe)
     if outputname != ""
         res |> CSV.write(outputname)
     end
@@ -68,12 +60,8 @@ function aggregatedoutput(inputname::AbstractString,outputname::AbstractString="
     isfile(inputname) || error("input file name does not exist")
     csvreader = CSVDateValReader(Dict(:filename=>inputname,:dateformat=>datefmt))
     valgator = DateValgator(COMMONARG)
-    lpipe = Pipeline(Dict(
-        :transformers => [csvreader,valgator]
-      )
-    )
-    fit!(lpipe)
-    res=transform!(lpipe)
+    lpipe = @pipeline csvreader |> valgator
+    res=fit_transform!(lpipe)
     if outputname != ""
         res |> CSV.write(outputname)
     end
@@ -87,12 +75,8 @@ function imputedstat(inputname::AbstractString,outputname::AbstractString="",dat
     valnner = DateValNNer(COMMONARG)
     mononicer = Monotonicer()
     stfier = Statifier()
-    lpipe = Pipeline(Dict(
-        :transformers => [csvreader,valgator,valnner,mononicer,stfier]
-      )
-    )
-    fit!(lpipe)
-    res=transform!(lpipe)
+    lpipe = @pipeline csvreader |> valgator |> valnner |> mononicer |> stfier
+    res=fit_transform!(lpipe)
     if outputname != ""
         res |> CSV.write(outputname)
     end
@@ -106,12 +90,8 @@ function imputedoutput(inputname::AbstractString,outputname::AbstractString="",d
     valnner = DateValNNer(COMMONARG)
     mononicer = Monotonicer()
     stfier = Statifier()
-    lpipe = Pipeline(Dict(
-        :transformers => [csvreader,valgator,valnner,mononicer]
-      )
-    )
-    fit!(lpipe)
-    res=transform!(lpipe)
+    lpipe = @pipeline csvreader |> valgator |> valnner |> mononicer
+    res=fit_transform!(lpipe)
     if outputname != ""
         res |> CSV.write(outputname)
     end
