+++ src/schema.jl
@@ -5,10 +5,9 @@ using Dates
 using DataFrames
 using JuliaDB: ML, table
 
-using TSML.TSMLTypes
-import TSML.TSMLTypes.fit! # to overload
-import TSML.TSMLTypes.transform! # to overload
-using TSML.Utils
+using AutoMLPipeline.AbsTypes
+using AutoMLPipeline.Utils
+import AutoMLPipeline.AbsTypes: fit!, transform!
 
 export fit!,transform!
 export Schemalizer, ML, table
