+++ test/test_timescaledb.jl
@@ -5,15 +5,14 @@ using TSML
 
 function test_timescaledb()
   tsdb = TimescaleDB()
-  mpipeline = Pipeline(Dict(
-         :transformers => [tsdb]
-     )
-  )
-  fit!(mpipeline)
-  respipe = transform!(mpipeline)
-  @test sum(respipe.Value) ≈ 560997.0
-  @test nrow(respipe) == 48349
-  @test ncol(respipe) == 2
+  mpipeline = @pipeline tsdb
+  try 
+    respipe = fit_transform!(mpipeline)
+    @test sum(respipe.Value) ≈ 560997.0
+    @test nrow(respipe) == 48349
+    @test ncol(respipe) == 2
+  catch e
+  end
 end
 @testset "TimescaleDB" begin
   test_timescaledb()
