+++ src/timescaledb.jl
@@ -4,10 +4,9 @@ using HTTP, JSON2
 using Dates
 using DataFrames
 
-using TSML.TSMLTypes
-import TSML.TSMLTypes.fit! # to overload
-import TSML.TSMLTypes.transform! # to overload
-using TSML.Utils
+using AutoMLPipeline.AbsTypes
+using AutoMLPipeline.Utils
+import AutoMLPipeline.AbsTypes: fit!, transform!
 
 export fit!,transform!
 export TimescaleDB

[…]
