+++ src/pkgdeps.jl
@@ -7,7 +7,7 @@ export mean, median, middle, quantile, std, var, cor, cov
 
 module PkgDeps
 
-using CSV: read, write
+import CSV: read, write
 using Dates: DateTime, Date,Dates, Day,Year, Month, Week, Minute, Hour, year, day, month, hour, minute, week, dayofweek, dayofquarter
 using DataFrames: DataFrame, DataFrameRow, DataFrameRows, DataFrameColumns, nrow, ncol, eachcol
 using Random: seed!
