+++ docs/src/lib/functions.md
@@ -6,22 +6,52 @@ Author = "Paulito Palmes"
 
 ## Index
 ```@index
-Modules = [TSML.DecisionTreeLearners,
-   TSML.Outliernicers, TSML.Plotters,
-   TSML.BaselineAlgos,TSML.EnsembleMethods,
-   TSML.MLBaseWrapper,TSML.Monotonicers,
-   TSML.Statifiers,TSML.BaseFilters,TSML.ValDateFilters,
+Modules = [
+   TSML.ArgumentParsers,
+   TSML.BaseFilters,
+   TSML.BaselineAlgos,
+   TSML.BaselineAlgos, 
+   TSML.CLIWrappers,
+   TSML.CrossValidators,
+   TSML.DecisionTreeLearners,
+   TSML.EnsembleMethods,
+   TSML.FeatureSelectors,
+   TSML.MLBaseWrapper,
+   TSML.Monotonicers,
+   TSML.Normalizers,
+   TSML.Outliernicers, 
+   TSML.Pipelines,
+   TSML.Plotters,
+   TSML.Schemalizers,
+   TSML.Statifiers,
    TSML.TSClassifiers,
-   TSML.TSMLTypes]
+   TSML.TSMLDemo,
+   TSML.ValDateFilters
+   ]
 ```
 
 ## Descriptions
 ```@autodocs
-Modules = [TSML.DecisionTreeLearners,
-   TSML.Outliernicers, TSML.Plotters,
-   TSML.BaselineAlgos,TSML.EnsembleMethods,
-   TSML.MLBaseWrapper,TSML.Monotonicers,
-   TSML.Statifiers,TSML.BaseFilters,TSML.ValDateFilters,
+Modules = [
+   TSML.ArgumentParsers,
+   TSML.BaseFilters,
+   TSML.BaselineAlgos,
+   TSML.BaselineAlgos, 
+   TSML.CLIWrappers,
+   TSML.CrossValidators,
+   TSML.DecisionTreeLearners,
+   TSML.EnsembleMethods,
+   TSML.FeatureSelectors,
+   TSML.MLBaseWrapper,
+   TSML.Monotonicers,
+   TSML.Normalizers,
+   TSML.Outliernicers, 
+   TSML.Pipelines,
+   TSML.Plotters,
+   TSML.Schemalizers,
+   TSML.Statifiers,
    TSML.TSClassifiers,
-   TSML.TSMLTypes]
+   TSML.TSMLDemo,
+   TSML.ValDateFilters
+   ]
 ```
