+++ .travis.yml
@@ -3,13 +3,19 @@ language: julia
 os:
   - linux
   - osx
+  - windows
 julia:
   - 1.0
   - 1.3
   - nightly
+addons:
+  apt:
+    packages:
+      - python3-numpy
+      - python3-scipy
 env:
   - JULIA_NUM_THREADS=8
-
+  - PYTHON=Conda
 notifications:
   email: false
 
@@ -22,6 +28,8 @@ after_success:
 jobs:
   allow_failures:
     - julia: nightly
+    - julia: 1.0
+    - os: windows
   include:
     - stage: "Documentation"
       julia: 1.3
