### 2020-03-11 4ab19e3
CompatHelper: add new compat entry for "AutoMLPipeline" at version "0.1"
