+++ Project.toml
@@ -8,7 +8,6 @@ AbstractTrees = "1520ce14-60c1-5f80-bbc7-55ef81b5835c"
 ArgParse = "c7e460c6-2fb9-53a9-8c5b-16f535851c63"
 AutoMLPipeline = "08437348-eef5-4817-bc1b-d4e9459680d6"
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
-CodecBzip2 = "523fee87-0ab8-5b00-afb7-3ecf72e48cfd"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 DecisionTree = "7806a523-6efd-50cb-b5f6-3fa6f1930dbb"
@@ -33,7 +32,6 @@ Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 AbstractTrees = "0.2, 0.3"
 ArgParse = "0.5, 0.6, 0.7, 1.0, 1.1"
 CSV = "0.4, 0.5"
-CodecBzip2 = "0.4, 0.5, 0.6"
 DataFrames = "0.17, 0.18, 0.19, 0.20"
 DecisionTree = "0.8, 0.9, 0.10"
 FileIO = "1.0, 1.1"
