### 2020-03-11 f9a9edf
docs and removal of pkgdeps export issue
