+++ versions.js
@@ -1,5 +1,6 @@
 var DOC_VERSIONS = [
   "stable",
+  "v2.5",
   "v2.4",
   "v2.3",
   "v2.1",
