### 2020-03-26 3f504e1
CompatHelper: bump compat for "StatsBase" to "0.33"
