### 2020-04-05 456c81d
CompatHelper: bump compat for "Plots" to "1.0"
