### 2020-04-24 f0381cf
CompatHelper: bump compat for "Plots" to "1.1"
