+++ src/TSML.jl
@@ -33,6 +33,8 @@ export Matrifier,Dateifier,
        DateValizer,DateValgator,DateValNNer,DateValMultiNNer,
        CSVDateValReader, CSVDateValWriter, DateValLinearImputer
        #BzCSVDateValReader
+export impute, impute!, chain, chain!, drop, drop!, 
+       interp, interp!, ImputeError, locf, nocb
 
 include("statifier.jl")
 using .Statifiers
