### 2020-04-29 97f86c5
remove test failure exception for 1.0
