+++ .travis.yml
@@ -27,7 +27,6 @@ after_success:
 jobs:
   allow_failures:
     - julia: nightly
-    - julia: 1.0
     - os: windows
   include:
     - stage: "Documentation"
