### 2020-04-30 294046e
CompatHelper: bump compat for "GR" to "0.49"
