+++ Project.toml
@@ -33,7 +33,7 @@ AbstractTrees = "0.2, 0.3"
 ArgParse = "0.5, 0.6, 0.7, 1.0, 1.1"
 AutoMLPipeline = "0.1"
 CSV = "0.4, 0.5, 0.6"
-DataFrames = "0.17, 0.18, 0.19, 0.20"
+DataFrames = "0.17, 0.18, 0.19, 0.20, 0.21"
 DecisionTree = "0.8, 0.9, 0.10"
 FileIO = "1.0, 1.1"
 GR = "0.35, 0.36, 0.37, 0.38, 0.39, 0.40, 0.41, 0.42, 0.43, 0.44, 0.45, 0.46, 0.47, 0.48"
