### 2020-05-29 2b55b28
CompatHelper: bump compat for "GR" to "0.50"
