+++ test/test_statifier.jl
@@ -13,13 +13,14 @@ function test_statifier()
   fit!(statfier,dat)
   res=transform!(statfier,dat)
   @test ncol(res) == 20
-  @test res[1,3:end]|>Vector|>sum |> x->round(x,digits=5) == 16.29114
+  @test res[1,3:end]|>Vector|>sum |> x->round(x,digits=5) == 7.74157
 
+  Random.seed!(123)
   statfier = Statifier(Dict(:processmissing=>true))
   fit!(statfier,dat)
   res=transform!(statfier,dat)
   @test ncol(res) == 26
-  @test res[1,3:end]|>Vector|>sum |> x->round(x,digits=5) == 28.29114
+  @test res[1,3:end]|>Vector|>sum |> x->round(x,digits=5) == 19.74157
 
   fname = joinpath(dirname(pathof(TSML)),"../data/testdata.csv")
   csvfilter = CSVDateValReader(Dict(:filename=>fname,:dateformat=>"dd/mm/yyyy HH:MM"))
