+++ Project.toml
@@ -33,7 +33,6 @@ AbstractTrees = "0.2, 0.3"
 ArgParse = "0.5, 0.6, 0.7, 1.0, 1.1"
 AutoMLPipeline = "0.1"
 CSV = "0.4, 0.5, 0.6"
-DataFrames = "0.17, 0.18, 0.19, 0.20"
 DecisionTree = "0.8, 0.9, 0.10"
 FileIO = "1.0, 1.1"
 GR = "0.35, 0.36, 0.37, 0.38, 0.39, 0.40, 0.41, 0.42, 0.43, 0.44, 0.45, 0.46, 0.47, 0.48"
@@ -42,7 +41,7 @@ IterTools = "1.1, 1.2, 1.3"
 JSON2 = "0.2, 0.3"
 JuliaDB = "0.13"
 MLBase = "0.6, 0.7, 0.8, 0.9"
-MLDataUtils = "0.2, 0.3, 0.4, 0.5"
+MLDataUtils = "0.2, 0.3, 0.4, 0.5, 0.6"
 MultivariateStats = "0.5, 0.6, 0.7, 0.8, 0.9"
 Plots = "0.25, 0.26, 0.27, 0.28, 0.29, 1.0, 1.1"
 StatsBase = "0.25, 0.26, 0.27, 0.28, 0.29, 0.30, 0.31, 0.32, 0.33"
