### 2020-07-05 4a585e4
CompatHelper: add new compat entry for "DataFrames" at version "0.21"
