### 2020-07-05 ccb3555
bug fixes in test
