### 2020-08-12 #51 @matbesancon
@ppalmes can you put the current release of the repo in a zenodo archive (zenodo.org). There is a direct link to Github and it should be quick to set up
