### 2020-08-12 #51 @ppalmes
@matbesancon , i cannot find my repo in the list for zenodo to authorize access because my repo is under IBM org and it seems the third party apps are not allowed. is there another way to use zenodo without github hooks? like uploading the repo itself? i'll ask the IBM org admin if they can add zenodo oauth.
