### 2020-08-12 #51 @matbesancon
yes you can also create the zenodo archive from scratch. What must be uploaded is the software package, not the paper
