### 2020-08-12 #51 @matbesancon
you should download the last version of the repo (like the last github release) and upload that to Zenodo
