### 2020-08-12 #51 @matbesancon
@ppalmes can you verify the final paper above? If ok, I'll close the issue, it is published :)
